+++ web/service/client.go
@@ -1,27 +1,21 @@
+// Package service implements the panel's business-logic layer.
+//
+// ClientService owns the lifecycle of VPN clients: creation, update, deletion,
+// attach/detach to inbounds, bulk operations, group membership, traffic resets,
+// and the paginated clients listing. Its surface is split across client_*.go
+// files by responsibility (see each file's contents); they all belong to the
+// same package, so the split is purely organizational. ClientService and
+// InboundService are mutually dependent — most ClientService methods take an
+// *InboundService and InboundService embeds a ClientService — which is why the
+// client code lives in package service rather than a sub-package.
 package service
 
 import (
-	"context"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
-	"fmt"
-	"slices"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
 
-	"github.com/google/uuid"
-	"github.com/mhsanaei/3x-ui/v3/database"
 	"github.com/mhsanaei/3x-ui/v3/database/model"
-	"github.com/mhsanaei/3x-ui/v3/logger"
-	"github.com/mhsanaei/3x-ui/v3/util/common"
-	"github.com/mhsanaei/3x-ui/v3/util/random"
 	"github.com/mhsanaei/3x-ui/v3/xray"
-
-	"gorm.io/gorm"
 )
 
 type ClientWithAttachments struct {
@@ -71,4384 +65,9 @@ type ClientService struct{}
 // between the clients table and the inbound's settings JSON.
 var ErrClientNotInInbound = errors.New("client not found in inbound")
 
-// Short-lived tombstone of just-deleted client emails so that a node snapshot
-// arriving between delete and node-side processing doesn't resurrect them.
-var (
-	recentlyDeletedMu sync.Mutex
-	recentlyDeleted   = map[string]time.Time{}
-)
-
-const deleteTombstoneTTL = 90 * time.Second
-
-var (
-	inboundMutationLocksMu sync.Mutex
-	inboundMutationLocks   = map[int]*sync.Mutex{}
-)
-
-func lockInbound(inboundId int) *sync.Mutex {
-	inboundMutationLocksMu.Lock()
-	defer inboundMutationLocksMu.Unlock()
-	m, ok := inboundMutationLocks[inboundId]
-	if !ok {
-		m = &sync.Mutex{}
-		inboundMutationLocks[inboundId] = m
-	}
-	m.Lock()
-	return m
-}
-
-func compactOrphans(db *gorm.DB, clients []any) []any {
-	if len(clients) == 0 {
-		return clients
-	}
-	emails := make([]string, 0, len(clients))
-	for _, c := range clients {
-		cm, ok := c.(map[string]any)
-		if !ok {
-			continue
-		}
-		if e, _ := cm["email"].(string); e != "" {
-			emails = append(emails, e)
-		}
-	}
-	if len(emails) == 0 {
-		return clients
-	}
-	existing := make(map[string]struct{}, len(emails))
-	const orphanChunk = 400
-	for start := 0; start < len(emails); start += orphanChunk {
-		end := min(start+orphanChunk, len(emails))
-		var found []string
-		if err := db.Model(&model.ClientRecord{}).Where("email IN ?", emails[start:end]).Pluck("email", &found).Error; err != nil {
-			logger.Warning("compactOrphans pluck:", err)
-			return clients
-		}
-		for _, e := range found {
-			existing[e] = struct{}{}
-		}
-	}
-	if len(existing) == len(emails) {
-		return clients
-	}
-	out := make([]any, 0, len(existing))
-	for _, c := range clients {
-		cm, ok := c.(map[string]any)
-		if !ok {
-			out = append(out, c)
-			continue
-		}
-		e, _ := cm["email"].(string)
-		if e == "" {
-			out = append(out, c)
-			continue
-		}
-		if _, ok := existing[e]; ok {
-			out = append(out, c)
-		}
-	}
-	return out
-}
-
-func tombstoneClientEmail(email string) {
-	if email == "" {
-		return
-	}
-	recentlyDeletedMu.Lock()
-	defer recentlyDeletedMu.Unlock()
-	recentlyDeleted[email] = time.Now()
-	cutoff := time.Now().Add(-deleteTombstoneTTL)
-	for e, ts := range recentlyDeleted {
-		if ts.Before(cutoff) {
-			delete(recentlyDeleted, e)
-		}
-	}
-}
-
-func tombstoneClientEmails(emails []string) {
-	if len(emails) == 0 {
-		return
-	}
-	now := time.Now()
-	cutoff := now.Add(-deleteTombstoneTTL)
-	recentlyDeletedMu.Lock()
-	defer recentlyDeletedMu.Unlock()
-	for _, email := range emails {
-		if email != "" {
-			recentlyDeleted[email] = now
-		}
-	}
-	for e, ts := range recentlyDeleted {
-		if ts.Before(cutoff) {
-			delete(recentlyDeleted, e)
-		}
-	}
-}
-
-func isClientEmailTombstoned(email string) bool {
-	if email == "" {
-		return false
-	}
-	recentlyDeletedMu.Lock()
-	defer recentlyDeletedMu.Unlock()
-	ts, ok := recentlyDeleted[email]
-	if !ok {
-		return false
-	}
-	if time.Since(ts) > deleteTombstoneTTL {
-		delete(recentlyDeleted, email)
-		return false
-	}
-	return true
-}
-
-func (s *ClientService) SyncInbound(tx *gorm.DB, inboundId int, clients []model.Client) error {
-	if tx == nil {
-		tx = database.GetDB()
-	}
-
-	if err := tx.Where("inbound_id = ?", inboundId).Delete(&model.ClientInbound{}).Error; err != nil {
-		return err
-	}
-
-	emails := make([]string, 0, len(clients))
-	seen := make(map[string]struct{}, len(clients))
-	for i := range clients {
-		email := strings.TrimSpace(clients[i].Email)
-		if email == "" {
-			continue
-		}
-		if _, ok := seen[email]; ok {
-			continue
-		}
-		seen[email] = struct{}{}
-		emails = append(emails, email)
-	}
-
-	existing := make(map[string]*model.ClientRecord, len(emails))
-	const selectChunk = 400
-	for start := 0; start < len(emails); start += selectChunk {
-		end := min(start+selectChunk, len(emails))
-		var rows []model.ClientRecord
-		if err := tx.Where("email IN ?", emails[start:end]).Find(&rows).Error; err != nil {
-			return err
-		}
-		for i := range rows {
-			r := rows[i]
-			existing[r.Email] = &r
-		}
-	}
-
-	idByEmail := make(map[string]int, len(emails))
-	pending := make(map[string]*model.ClientRecord, len(emails))
-	toCreate := make([]*model.ClientRecord, 0, len(emails))
-	for i := range clients {
-		email := strings.TrimSpace(clients[i].Email)
-		if email == "" {
-			continue
-		}
-
-		incoming := clients[i].ToRecord()
-		row, ok := existing[email]
-		if !ok {
-			if _, dup := pending[email]; !dup {
-				pending[email] = incoming
-				toCreate = append(toCreate, incoming)
-			}
-			continue
-		}
-
-		before := *row
-		if incoming.UUID != "" {
-			row.UUID = incoming.UUID
-		}
-		if incoming.Password != "" {
-			row.Password = incoming.Password
-		}
-		if incoming.Auth != "" {
-			row.Auth = incoming.Auth
-		}
-		row.Flow = incoming.Flow
-		if incoming.Security != "" {
-			row.Security = incoming.Security
-		}
-		if incoming.Reverse != "" {
-			row.Reverse = incoming.Reverse
-		}
-		row.SubID = incoming.SubID
-		row.LimitIP = incoming.LimitIP
-		row.TotalGB = incoming.TotalGB
-		row.ExpiryTime = incoming.ExpiryTime
-		row.Enable = incoming.Enable
-		row.TgID = incoming.TgID
-		if incoming.Group != "" {
-			row.Group = incoming.Group
-		}
-		row.Comment = incoming.Comment
-		row.Reset = incoming.Reset
-		if incoming.CreatedAt > 0 && (row.CreatedAt == 0 || incoming.CreatedAt < row.CreatedAt) {
-			row.CreatedAt = incoming.CreatedAt
-		}
-		preservedUpdatedAt := max(incoming.UpdatedAt, row.UpdatedAt)
-		row.UpdatedAt = preservedUpdatedAt
-
-		idByEmail[email] = row.Id
-
-		if *row == before {
-			continue
-		}
-		if err := tx.Save(row).Error; err != nil {
-			return err
-		}
-		if err := tx.Model(&model.ClientRecord{}).
-			Where("id = ?", row.Id).
-			UpdateColumn("updated_at", preservedUpdatedAt).Error; err != nil {
-			return err
-		}
-	}
-
-	if len(toCreate) > 0 {
-		if err := tx.CreateInBatches(toCreate, 200).Error; err != nil {
-			return err
-		}
-		for _, rec := range toCreate {
-			idByEmail[rec.Email] = rec.Id
-		}
-	}
-
-	links := make([]model.ClientInbound, 0, len(clients))
-	linked := make(map[int]struct{}, len(clients))
-	for i := range clients {
-		email := strings.TrimSpace(clients[i].Email)
-		if email == "" {
-			continue
-		}
-		id, ok := idByEmail[email]
-		if !ok {
-			continue
-		}
-		if _, dup := linked[id]; dup {
-			continue
-		}
-		linked[id] = struct{}{}
-		links = append(links, model.ClientInbound{
-			ClientId:     id,
-			InboundId:    inboundId,
-			FlowOverride: clients[i].Flow,
-		})
-	}
-	if len(links) > 0 {
-		if err := tx.CreateInBatches(links, 200).Error; err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (s *ClientService) DetachInbound(tx *gorm.DB, inboundId int) error {
-	if tx == nil {
-		tx = database.GetDB()
-	}
-	return tx.Where("inbound_id = ?", inboundId).Delete(&model.ClientInbound{}).Error
-}
-
-func (s *ClientService) ListForInbound(tx *gorm.DB, inboundId int) ([]model.Client, error) {
-	if tx == nil {
-		tx = database.GetDB()
-	}
-	type joinedRow struct {
-		model.ClientRecord
-		FlowOverride string
-	}
-	var rows []joinedRow
-	err := tx.Table("clients").
-		Select("clients.*, client_inbounds.flow_override AS flow_override").
-		Joins("JOIN client_inbounds ON client_inbounds.client_id = clients.id").
-		Where("client_inbounds.inbound_id = ?", inboundId).
-		Order("clients.id ASC").
-		Find(&rows).Error
-	if err != nil {
-		return nil, err
-	}
-
-	out := make([]model.Client, 0, len(rows))
-	for i := range rows {
-		c := rows[i].ToClient()
-		c.Flow = rows[i].FlowOverride
-		out = append(out, *c)
-	}
-	return out, nil
-}
-
-func (s *ClientService) GetRecordByEmail(tx *gorm.DB, email string) (*model.ClientRecord, error) {
-	if tx == nil {
-		tx = database.GetDB()
-	}
-	row := &model.ClientRecord{}
-	err := tx.Where("email = ?", email).First(row).Error
-	if err != nil {
-		return nil, err
-	}
-	return row, nil
-}
-
-// EffectiveFlow returns the client's flow from the first flow-capable inbound
-// it is attached to (lowest inbound_id with a non-empty flow_override). The
-// canonical clients.Flow column is unreliable for multi-inbound clients: a
-// non-flow inbound (Hysteria, WS, gRPC, …) carries an empty flow and, when its
-// SyncInbound runs last, overwrites the column to "" even though a VLESS Reality
-// inbound stored a real flow. The per-inbound flow_override is always correct,
-// so derive the display flow from it (order-independent). See issue #4792.
-func (s *ClientService) EffectiveFlow(tx *gorm.DB, recordId int) (string, error) {
-	if tx == nil {
-		tx = database.GetDB()
-	}
-	var flows []string
-	err := tx.Model(&model.ClientInbound{}).
-		Where("client_id = ? AND flow_override <> ?", recordId, "").
-		Order("inbound_id ASC").
-		Limit(1).
-		Pluck("flow_override", &flows).Error
-	if err != nil {
-		return "", err
-	}
-	if len(flows) == 0 {
-		return "", nil
-	}
-	return flows[0], nil
-}
-
-func (s *ClientService) GetInboundIdsForEmail(tx *gorm.DB, email string) ([]int, error) {
-	if tx == nil {
-		tx = database.GetDB()
-	}
-	var ids []int
-	err := tx.Table("client_inbounds").
-		Select("client_inbounds.inbound_id").
-		Joins("JOIN clients ON clients.id = client_inbounds.client_id").
-		Where("clients.email = ?", email).
-		Scan(&ids).Error
-	if err != nil {
-		return nil, err
-	}
-	return ids, nil
-}
-
-func (s *ClientService) GetByID(id int) (*model.ClientRecord, error) {
-	row := &model.ClientRecord{}
-	if err := database.GetDB().Where("id = ?", id).First(row).Error; err != nil {
-		return nil, err
-	}
-	return row, nil
-}
-
-func (s *ClientService) GetInboundIdsForRecord(id int) ([]int, error) {
-	var ids []int
-	err := database.GetDB().Table("client_inbounds").
-		Where("client_id = ?", id).
-		Order("inbound_id ASC").
-		Pluck("inbound_id", &ids).Error
-	if err != nil {
-		return nil, err
-	}
-	return ids, nil
-}
-
-func (s *ClientService) List() ([]ClientWithAttachments, error) {
-	db := database.GetDB()
-	var rows []model.ClientRecord
-	if err := db.Order("id ASC").Find(&rows).Error; err != nil {
-		return nil, err
-	}
-	if len(rows) == 0 {
-		return []ClientWithAttachments{}, nil
-	}
-
-	clientIds := make([]int, 0, len(rows))
-	emails := make([]string, 0, len(rows))
-	for i := range rows {
-		clientIds = append(clientIds, rows[i].Id)
-		if rows[i].Email != "" {
-			emails = append(emails, rows[i].Email)
-		}
-	}
-
-	attachments := make(map[int][]int, len(rows))
-	for _, batch := range chunkInts(clientIds, sqlInChunk) {
-		var links []model.ClientInbound
-		if err := db.Where("client_id IN ?", batch).Find(&links).Error; err != nil {
-			return nil, err
-		}
-		for _, l := range links {
-			attachments[l.ClientId] = append(attachments[l.ClientId], l.InboundId)
-		}
-	}
-
-	trafficByEmail := make(map[string]*xray.ClientTraffic, len(emails))
-	if len(emails) > 0 {
-		var stats []xray.ClientTraffic
-		for _, batch := range chunkStrings(emails, sqlInChunk) {
-			var batchStats []xray.ClientTraffic
-			if err := db.Where("email IN ?", batch).Find(&batchStats).Error; err != nil {
-				return nil, err
-			}
-			stats = append(stats, batchStats...)
-		}
-		for i := range stats {
-			trafficByEmail[stats[i].Email] = &stats[i]
-		}
-	}
-
-	out := make([]ClientWithAttachments, 0, len(rows))
-	for i := range rows {
-		out = append(out, ClientWithAttachments{
-			ClientRecord: rows[i],
-			InboundIds:   attachments[rows[i].Id],
-			Traffic:      trafficByEmail[rows[i].Email],
-		})
-	}
-	return out, nil
-}
-
 type ClientCreatePayload struct {
 	Client     model.Client `json:"client"`
 	InboundIds []int        `json:"inboundIds"`
 }
 
-func hasForbiddenClientChar(s string) bool {
-	for _, r := range s {
-		if r == '/' || r == '\\' || r == ' ' || r < 0x20 || r == 0x7f {
-			return true
-		}
-	}
-	return false
-}
-
-func validateClientEmail(email string) error {
-	if hasForbiddenClientChar(email) {
-		return common.NewError("client email contains an invalid character:", email)
-	}
-	return nil
-}
-
-func validateClientSubID(subID string) error {
-	if hasForbiddenClientChar(subID) {
-		return common.NewError("client subId contains an invalid character:", subID)
-	}
-	return nil
-}
-
-func (s *ClientService) HasPendingNode(inboundSvc *InboundService, email string) bool {
-	if strings.TrimSpace(email) == "" {
-		return false
-	}
-	ids, err := s.GetInboundIdsForEmail(nil, email)
-	if err != nil {
-		return false
-	}
-	return inboundSvc.AnyNodePending(ids)
-}
-
-func (s *ClientService) Create(inboundSvc *InboundService, payload *ClientCreatePayload) (bool, error) {
-	if payload == nil {
-		return false, common.NewError("empty payload")
-	}
-	client := payload.Client
-	if strings.TrimSpace(client.Email) == "" {
-		return false, common.NewError("client email is required")
-	}
-	if err := validateClientEmail(client.Email); err != nil {
-		return false, err
-	}
-	if err := validateClientSubID(client.SubID); err != nil {
-		return false, err
-	}
-	if len(payload.InboundIds) == 0 {
-		return false, common.NewError("at least one inbound is required")
-	}
-
-	if client.SubID == "" {
-		client.SubID = uuid.NewString()
-	}
-	if !client.Enable {
-		client.Enable = true
-	}
-	now := time.Now().UnixMilli()
-	if client.CreatedAt == 0 {
-		client.CreatedAt = now
-	}
-	client.UpdatedAt = now
-
-	existing := &model.ClientRecord{}
-	err := database.GetDB().Where("email = ?", client.Email).First(existing).Error
-	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
-		return false, err
-	}
-	emailTaken := !errors.Is(err, gorm.ErrRecordNotFound)
-	if emailTaken {
-		if existing.SubID == "" || existing.SubID != client.SubID {
-			return false, common.NewError("email already in use:", client.Email)
-		}
-	}
-
-	if client.SubID != "" {
-		var subTaken int64
-		if err := database.GetDB().Model(&model.ClientRecord{}).
-			Where("sub_id = ? AND email <> ?", client.SubID, client.Email).
-			Count(&subTaken).Error; err != nil {
-			return false, err
-		}
-		if subTaken > 0 {
-			return false, common.NewError("subId already in use:", client.SubID)
-		}
-	}
-
-	needRestart := false
-	for _, ibId := range payload.InboundIds {
-		inbound, getErr := inboundSvc.GetInbound(ibId)
-		if getErr != nil {
-			return needRestart, getErr
-		}
-		if err := s.fillProtocolDefaults(&client, inbound); err != nil {
-			return needRestart, err
-		}
-		settingsPayload, mErr := json.Marshal(map[string][]model.Client{"clients": {clientWithInboundFlow(client, inbound)}})
-		if mErr != nil {
-			return needRestart, mErr
-		}
-		nr, addErr := s.AddInboundClient(inboundSvc, &model.Inbound{
-			Id:       ibId,
-			Settings: string(settingsPayload),
-		})
-		if addErr != nil {
-			return needRestart, addErr
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) fillProtocolDefaults(c *model.Client, ib *model.Inbound) error {
-	switch ib.Protocol {
-	case model.VMESS, model.VLESS:
-		if c.ID == "" {
-			c.ID = uuid.NewString()
-		}
-	case model.Trojan:
-		if c.Password == "" {
-			c.Password = strings.ReplaceAll(uuid.NewString(), "-", "")
-		}
-	case model.Shadowsocks:
-		method := shadowsocksMethodFromSettings(ib.Settings)
-		if c.Password == "" || !validShadowsocksClientKey(method, c.Password) {
-			c.Password = randomShadowsocksClientKey(method)
-		}
-	case model.Hysteria:
-		if c.Auth == "" {
-			c.Auth = strings.ReplaceAll(uuid.NewString(), "-", "")
-		}
-	}
-	return nil
-}
-
-func clientWithInboundFlow(c model.Client, ib *model.Inbound) model.Client {
-	if !inboundCanEnableTlsFlow(string(ib.Protocol), ib.StreamSettings) {
-		c.Flow = ""
-	}
-	return c
-}
-
-func shadowsocksMethodFromSettings(settings string) string {
-	if settings == "" {
-		return ""
-	}
-	var m map[string]any
-	if err := json.Unmarshal([]byte(settings), &m); err != nil {
-		return ""
-	}
-	method, _ := m["method"].(string)
-	return method
-}
-
-func randomShadowsocksClientKey(method string) string {
-	if n := shadowsocksKeyBytes(method); n > 0 {
-		return random.Base64Bytes(n)
-	}
-	return strings.ReplaceAll(uuid.NewString(), "-", "")
-}
-
-func validShadowsocksClientKey(method, key string) bool {
-	n := shadowsocksKeyBytes(method)
-	if n == 0 {
-		return key != ""
-	}
-	decoded, err := base64.StdEncoding.DecodeString(key)
-	if err != nil {
-		return false
-	}
-	return len(decoded) == n
-}
-
-func shadowsocksKeyBytes(method string) int {
-	switch method {
-	case "2022-blake3-aes-128-gcm":
-		return 16
-	case "2022-blake3-aes-256-gcm", "2022-blake3-chacha20-poly1305":
-		return 32
-	}
-	return 0
-}
-
-func applyShadowsocksClientMethod(clients []any, settings map[string]any) {
-	method, _ := settings["method"].(string)
-	is2022 := strings.HasPrefix(method, "2022-blake3-")
-	for i := range clients {
-		cm, ok := clients[i].(map[string]any)
-		if !ok {
-			continue
-		}
-		if is2022 {
-			if _, hasKey := cm["method"]; hasKey {
-				delete(cm, "method")
-				clients[i] = cm
-			}
-			continue
-		}
-		if method == "" {
-			continue
-		}
-		if existing, _ := cm["method"].(string); existing != "" {
-			continue
-		}
-		cm["method"] = method
-		clients[i] = cm
-	}
-}
-
-func (s *ClientService) Update(inboundSvc *InboundService, id int, updated model.Client, inboundFilter ...int) (bool, error) {
-	existing, err := s.GetByID(id)
-	if err != nil {
-		return false, err
-	}
-	inboundIds, err := s.GetInboundIdsForRecord(id)
-	if err != nil {
-		return false, err
-	}
-	if len(inboundFilter) > 0 {
-		allow := make(map[int]struct{}, len(inboundFilter))
-		for _, fid := range inboundFilter {
-			allow[fid] = struct{}{}
-		}
-		filtered := inboundIds[:0:0]
-		for _, ibId := range inboundIds {
-			if _, ok := allow[ibId]; ok {
-				filtered = append(filtered, ibId)
-			}
-		}
-		inboundIds = filtered
-	}
-
-	if strings.TrimSpace(updated.Email) == "" {
-		return false, common.NewError("client email is required")
-	}
-	if err := validateClientEmail(updated.Email); err != nil {
-		return false, err
-	}
-	if err := validateClientSubID(updated.SubID); err != nil {
-		return false, err
-	}
-	if updated.SubID == "" {
-		updated.SubID = existing.SubID
-	}
-	if updated.SubID == "" {
-		updated.SubID = uuid.NewString()
-	}
-	updated.UpdatedAt = time.Now().UnixMilli()
-	if updated.CreatedAt == 0 {
-		updated.CreatedAt = existing.CreatedAt
-	}
-
-	// Preserve existing credentials when the caller omits them, so a partial
-	// update (e.g. only changing traffic/expiry) doesn't silently rotate the
-	// client's UUID/password/auth via fillProtocolDefaults. Supplying a new
-	// value still rotates it intentionally.
-	if updated.ID == "" {
-		updated.ID = existing.UUID
-	}
-	if updated.Password == "" {
-		updated.Password = existing.Password
-	}
-	if updated.Auth == "" {
-		updated.Auth = existing.Auth
-	}
-
-	if updated.Email != existing.Email {
-		var collisionCount int64
-		if err := database.GetDB().Model(&model.ClientRecord{}).
-			Where("email = ? AND id <> ?", updated.Email, id).
-			Count(&collisionCount).Error; err != nil {
-			return false, err
-		}
-		if collisionCount > 0 {
-			return false, common.NewError("Duplicate email:", updated.Email)
-		}
-		if err := database.GetDB().Model(&model.ClientRecord{}).
-			Where("id = ?", id).
-			Update("email", updated.Email).Error; err != nil {
-			return false, err
-		}
-	}
-
-	if updated.SubID != "" {
-		var subCollision int64
-		if err := database.GetDB().Model(&model.ClientRecord{}).
-			Where("sub_id = ? AND id <> ?", updated.SubID, id).
-			Count(&subCollision).Error; err != nil {
-			return false, err
-		}
-		if subCollision > 0 {
-			return false, common.NewError("Duplicate subId:", updated.SubID)
-		}
-	}
-
-	needRestart := false
-	for _, ibId := range inboundIds {
-		inbound, getErr := inboundSvc.GetInbound(ibId)
-		if getErr != nil {
-			if errors.Is(getErr, gorm.ErrRecordNotFound) {
-				if err := database.GetDB().
-					Where("client_id = ? AND inbound_id = ?", id, ibId).
-					Delete(&model.ClientInbound{}).Error; err != nil {
-					return needRestart, err
-				}
-				continue
-			}
-			return needRestart, getErr
-		}
-		if existing.Email == "" {
-			continue
-		}
-		if err := s.fillProtocolDefaults(&updated, inbound); err != nil {
-			return needRestart, err
-		}
-		settingsPayload, mErr := json.Marshal(map[string][]model.Client{"clients": {clientWithInboundFlow(updated, inbound)}})
-		if mErr != nil {
-			return needRestart, mErr
-		}
-		nr, upErr := s.UpdateInboundClient(inboundSvc, &model.Inbound{
-			Id:       ibId,
-			Settings: string(settingsPayload),
-		}, existing.Email)
-		if upErr != nil {
-			return needRestart, upErr
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-
-	reverseStr := ""
-	if updated.Reverse != nil && strings.TrimSpace(updated.Reverse.Tag) != "" {
-		if b, mErr := json.Marshal(updated.Reverse); mErr == nil {
-			reverseStr = string(b)
-		}
-	}
-	if err := database.GetDB().Model(&model.ClientRecord{}).
-		Where("id = ?", id).
-		Update("reverse", reverseStr).Error; err != nil {
-		return needRestart, err
-	}
-
-	if err := database.GetDB().Model(&model.ClientRecord{}).
-		Where("id = ?", id).
-		UpdateColumn("updated_at", time.Now().UnixMilli()).Error; err != nil {
-		return needRestart, err
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) Delete(inboundSvc *InboundService, id int, keepTraffic bool) (bool, error) {
-	existing, err := s.GetByID(id)
-	if err != nil {
-		return false, err
-	}
-	tombstoneClientEmail(existing.Email)
-
-	inboundIds, err := s.GetInboundIdsForRecord(id)
-	if err != nil {
-		return false, err
-	}
-
-	needRestart := false
-	for _, ibId := range inboundIds {
-		if _, getErr := inboundSvc.GetInbound(ibId); getErr != nil {
-			if errors.Is(getErr, gorm.ErrRecordNotFound) {
-				continue
-			}
-			return needRestart, getErr
-		}
-
-		// Always delete by email — the client's stable identity. This removes
-		// every matching entry from the inbound's settings even when the stored
-		// credential (UUID/password/auth) drifted from the inbound JSON, or a
-		// duplicate entry with the same email exists.
-		if existing.Email == "" {
-			continue
-		}
-		nr, delErr := s.DelInboundClientByEmail(inboundSvc, ibId, existing.Email, false)
-		if delErr != nil {
-			// The client is already absent from this inbound (data drift or a
-			// retried delete). Skip it — deletion stays idempotent.
-			if errors.Is(delErr, ErrClientNotInInbound) {
-				continue
-			}
-			return needRestart, delErr
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-
-	db := database.GetDB()
-	if err := db.Where("client_id = ?", id).Delete(&model.ClientInbound{}).Error; err != nil {
-		return needRestart, err
-	}
-	if !keepTraffic && existing.Email != "" {
-		if err := db.Where("email = ?", existing.Email).Delete(&xray.ClientTraffic{}).Error; err != nil {
-			return needRestart, err
-		}
-		if err := db.Where("client_email = ?", existing.Email).Delete(&model.InboundClientIps{}).Error; err != nil {
-			return needRestart, err
-		}
-	}
-	if err := db.Delete(&model.ClientRecord{}, id).Error; err != nil {
-		return needRestart, err
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) Attach(inboundSvc *InboundService, id int, inboundIds []int) (bool, error) {
-	existing, err := s.GetByID(id)
-	if err != nil {
-		return false, err
-	}
-	currentIds, err := s.GetInboundIdsForRecord(id)
-	if err != nil {
-		return false, err
-	}
-	have := make(map[int]struct{}, len(currentIds))
-	for _, x := range currentIds {
-		have[x] = struct{}{}
-	}
-
-	clientWire := existing.ToClient()
-	flow, ffErr := s.EffectiveFlow(nil, id)
-	if ffErr != nil {
-		return false, ffErr
-	}
-	clientWire.Flow = flow
-	clientWire.UpdatedAt = time.Now().UnixMilli()
-
-	needRestart := false
-	for _, ibId := range inboundIds {
-		if _, attached := have[ibId]; attached {
-			continue
-		}
-		inbound, getErr := inboundSvc.GetInbound(ibId)
-		if getErr != nil {
-			return needRestart, getErr
-		}
-		copyClient := *clientWire
-		if err := s.fillProtocolDefaults(&copyClient, inbound); err != nil {
-			return needRestart, err
-		}
-		settingsPayload, mErr := json.Marshal(map[string][]model.Client{"clients": {clientWithInboundFlow(copyClient, inbound)}})
-		if mErr != nil {
-			return needRestart, mErr
-		}
-		nr, addErr := s.AddInboundClient(inboundSvc, &model.Inbound{
-			Id:       ibId,
-			Settings: string(settingsPayload),
-		})
-		if addErr != nil {
-			return needRestart, addErr
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) CreateOne(inboundSvc *InboundService, inboundId int, client model.Client) (bool, error) {
-	return s.Create(inboundSvc, &ClientCreatePayload{
-		Client:     client,
-		InboundIds: []int{inboundId},
-	})
-}
-
-func (s *ClientService) DetachByEmail(inboundSvc *InboundService, inboundId int, email string) (bool, error) {
-	if email == "" {
-		return false, common.NewError("client email is required")
-	}
-	rec, err := s.GetRecordByEmail(nil, email)
-	if err != nil {
-		return false, err
-	}
-	return s.Detach(inboundSvc, rec.Id, []int{inboundId})
-}
-
-func (s *ClientService) AttachByEmail(inboundSvc *InboundService, email string, inboundIds []int) (bool, error) {
-	if email == "" {
-		return false, common.NewError("client email is required")
-	}
-	rec, err := s.GetRecordByEmail(nil, email)
-	if err != nil {
-		return false, err
-	}
-	return s.Attach(inboundSvc, rec.Id, inboundIds)
-}
-
-// BulkAttachResult reports the outcome of a bulk attach across target inbounds.
-type BulkAttachResult struct {
-	Attached []string `json:"attached"`
-	Skipped  []string `json:"skipped"`
-	Errors   []string `json:"errors"`
-}
-
-// BulkAttach attaches the given existing clients (by email) to each target inbound,
-// reusing their identity (email/UUID/password/subId) and a shared traffic row. It adds
-// all clients to a target in a single AddInboundClient call, and reports clients already
-// present on a target as skipped.
-func (s *ClientService) BulkAttach(inboundSvc *InboundService, emails []string, inboundIds []int) (*BulkAttachResult, bool, error) {
-	result := &BulkAttachResult{}
-	if len(emails) == 0 || len(inboundIds) == 0 {
-		return result, false, nil
-	}
-
-	recordErr := func(format string, args ...any) {
-		msg := fmt.Sprintf(format, args...)
-		result.Errors = append(result.Errors, msg)
-		logger.Warningf("[BulkAttach] %s", msg)
-	}
-
-	records := make([]*model.ClientRecord, 0, len(emails))
-	seenEmail := make(map[string]struct{}, len(emails))
-	for _, email := range emails {
-		if email == "" {
-			continue
-		}
-		key := strings.ToLower(email)
-		if _, ok := seenEmail[key]; ok {
-			continue
-		}
-		seenEmail[key] = struct{}{}
-		rec, err := s.GetRecordByEmail(nil, email)
-		if err != nil {
-			recordErr("%s: %v", email, err)
-			continue
-		}
-		records = append(records, rec)
-	}
-
-	emailSubIDs, sidErr := inboundSvc.getAllEmailSubIDs()
-	if sidErr != nil {
-		emailSubIDs = nil
-		logger.Warningf("[BulkAttach] getAllEmailSubIDs: %v", sidErr)
-	}
-
-	needRestart := false
-	for _, ibId := range inboundIds {
-		inbound, err := inboundSvc.GetInbound(ibId)
-		if err != nil {
-			recordErr("inbound %d: %v", ibId, err)
-			continue
-		}
-		existingClients, err := inboundSvc.GetClients(inbound)
-		if err != nil {
-			recordErr("inbound %d: %v", ibId, err)
-			continue
-		}
-		have := make(map[string]struct{}, len(existingClients))
-		for _, c := range existingClients {
-			have[strings.ToLower(c.Email)] = struct{}{}
-		}
-
-		clientsToAdd := make([]model.Client, 0, len(records))
-		for _, rec := range records {
-			if _, attached := have[strings.ToLower(rec.Email)]; attached {
-				result.Skipped = append(result.Skipped, rec.Email)
-				continue
-			}
-			client := *rec.ToClient()
-			client.UpdatedAt = time.Now().UnixMilli()
-			if err := s.fillProtocolDefaults(&client, inbound); err != nil {
-				recordErr("%s -> inbound %d: %v", rec.Email, ibId, err)
-				continue
-			}
-			clientsToAdd = append(clientsToAdd, clientWithInboundFlow(client, inbound))
-		}
-
-		if len(clientsToAdd) == 0 {
-			continue
-		}
-
-		payload, err := json.Marshal(map[string][]model.Client{"clients": clientsToAdd})
-		if err != nil {
-			recordErr("inbound %d: %v", ibId, err)
-			continue
-		}
-		nr, err := s.addInboundClient(inboundSvc, &model.Inbound{Id: ibId, Settings: string(payload)}, emailSubIDs)
-		if err != nil {
-			recordErr("inbound %d: %v", ibId, err)
-			continue
-		}
-		if nr {
-			needRestart = true
-		}
-		for _, c := range clientsToAdd {
-			result.Attached = append(result.Attached, c.Email)
-		}
-	}
-
-	return result, needRestart, nil
-}
-
-// BulkDetachResult reports the outcome of a bulk detach across target inbounds.
-type BulkDetachResult struct {
-	Detached []string `json:"detached"`
-	Skipped  []string `json:"skipped"`
-	Errors   []string `json:"errors"`
-}
-
-// BulkDetach detaches the given existing clients (by email) from each target inbound.
-// (email, inbound) pairs where the client is not currently attached are silently skipped
-// at the inbound level; emails that aren't attached to any of the requested inbounds
-// are reported under skipped. ClientRecord rows are kept even when they become orphaned
-// (matches single-client detach semantics); callers should use bulkDelete for full removal.
-func (s *ClientService) BulkDetach(inboundSvc *InboundService, emails []string, inboundIds []int) (*BulkDetachResult, bool, error) {
-	result := &BulkDetachResult{}
-	if len(emails) == 0 || len(inboundIds) == 0 {
-		return result, false, nil
-	}
-
-	recordErr := func(format string, args ...any) {
-		msg := fmt.Sprintf(format, args...)
-		result.Errors = append(result.Errors, msg)
-		logger.Warningf("[BulkDetach] %s", msg)
-	}
-
-	requested := make(map[int]struct{}, len(inboundIds))
-	for _, id := range inboundIds {
-		requested[id] = struct{}{}
-	}
-
-	recsByInbound := make(map[int][]*model.ClientRecord)
-	emailOrder := make([]string, 0, len(emails))
-	emailRepr := make(map[string]string, len(emails))
-	emailFailed := make(map[string]bool, len(emails))
-	seenEmail := make(map[string]struct{}, len(emails))
-	for _, email := range emails {
-		if email == "" {
-			continue
-		}
-		key := strings.ToLower(email)
-		if _, ok := seenEmail[key]; ok {
-			continue
-		}
-		seenEmail[key] = struct{}{}
-
-		rec, err := s.GetRecordByEmail(nil, email)
-		if err != nil {
-			recordErr("%s: %v", email, err)
-			continue
-		}
-		currentIds, err := s.GetInboundIdsForRecord(rec.Id)
-		if err != nil {
-			recordErr("%s: %v", email, err)
-			continue
-		}
-		matched := false
-		for _, id := range currentIds {
-			if _, ok := requested[id]; ok {
-				recsByInbound[id] = append(recsByInbound[id], rec)
-				matched = true
-			}
-		}
-		if !matched {
-			result.Skipped = append(result.Skipped, rec.Email)
-			continue
-		}
-		emailOrder = append(emailOrder, key)
-		emailRepr[key] = rec.Email
-	}
-
-	needRestart := false
-	for _, ibId := range inboundIds {
-		recs, ok := recsByInbound[ibId]
-		if !ok {
-			continue
-		}
-		delete(recsByInbound, ibId)
-		nr, err := s.delInboundClients(inboundSvc, ibId, recs, true)
-		if err != nil {
-			recordErr("inbound %d: %v", ibId, err)
-			for _, rec := range recs {
-				emailFailed[strings.ToLower(rec.Email)] = true
-			}
-			continue
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-
-	for _, key := range emailOrder {
-		if emailFailed[key] {
-			continue
-		}
-		result.Detached = append(result.Detached, emailRepr[key])
-	}
-
-	return result, needRestart, nil
-}
-
-// delInboundClients removes several clients from a single inbound in one pass:
-// one settings rewrite, one runtime sweep, one Save and one SyncInbound for the
-// whole batch, instead of repeating the full per-client cycle. It mirrors the
-// semantics of DelInboundClientByEmail for each removed client. needRestart is
-// the OR across all removals.
-func (s *ClientService) delInboundClients(inboundSvc *InboundService, inboundId int, recs []*model.ClientRecord, keepTraffic bool) (bool, error) {
-	if len(recs) == 0 {
-		return false, nil
-	}
-	defer lockInbound(inboundId).Unlock()
-
-	oldInbound, err := inboundSvc.GetInbound(inboundId)
-	if err != nil {
-		logger.Error("Load Old Data Error")
-		return false, err
-	}
-
-	var settings map[string]any
-	if err := json.Unmarshal([]byte(oldInbound.Settings), &settings); err != nil {
-		return false, err
-	}
-
-	// Match by email — the client's stable identity (see Delete). Removes every
-	// entry carrying a wanted email, independent of credential drift.
-	wanted := make(map[string]struct{}, len(recs))
-	for _, rec := range recs {
-		if rec.Email != "" {
-			wanted[rec.Email] = struct{}{}
-		}
-	}
-
-	interfaceClients, ok := settings["clients"].([]any)
-	if !ok {
-		return false, common.NewError("invalid clients format in inbound settings")
-	}
-
-	type removedClient struct {
-		email      string
-		needApiDel bool
-	}
-	removed := make([]removedClient, 0, len(wanted))
-	newClients := make([]any, 0, len(interfaceClients))
-	for _, client := range interfaceClients {
-		c, ok := client.(map[string]any)
-		if !ok {
-			newClients = append(newClients, client)
-			continue
-		}
-		email, _ := c["email"].(string)
-		if _, hit := wanted[email]; hit && email != "" {
-			enable, _ := c["enable"].(bool)
-			removed = append(removed, removedClient{email: email, needApiDel: enable})
-			continue
-		}
-		newClients = append(newClients, client)
-	}
-
-	if len(removed) == 0 {
-		return false, nil
-	}
-
-	db := database.GetDB()
-	newClients = compactOrphans(db, newClients)
-	if newClients == nil {
-		newClients = []any{}
-	}
-	settings["clients"] = newClients
-	newSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-	oldInbound.Settings = string(newSettings)
-
-	var sharedSet map[string]bool
-	if !keepTraffic {
-		removedEmails := make([]string, 0, len(removed))
-		for _, r := range removed {
-			if r.email != "" {
-				removedEmails = append(removedEmails, r.email)
-			}
-		}
-		var sharedErr error
-		sharedSet, sharedErr = inboundSvc.emailsUsedByOtherInbounds(removedEmails, inboundId)
-		if sharedErr != nil {
-			return false, sharedErr
-		}
-	}
-
-	needRestart := false
-	markDirty := false
-	for _, r := range removed {
-		email := r.email
-		emailShared := sharedSet[strings.ToLower(strings.TrimSpace(email))]
-		if !emailShared && !keepTraffic {
-			if err := inboundSvc.DelClientIPs(db, email); err != nil {
-				logger.Error("Error in delete client IPs")
-				return needRestart, err
-			}
-		}
-		if len(email) > 0 {
-			var enables []bool
-			if err := db.Model(xray.ClientTraffic{}).Where("email = ?", email).Limit(1).Pluck("enable", &enables).Error; err != nil {
-				logger.Error("Get stats error")
-				return needRestart, err
-			}
-			notDepleted := len(enables) > 0 && enables[0]
-			if !emailShared && !keepTraffic {
-				if err := inboundSvc.DelClientStat(db, email); err != nil {
-					logger.Error("Delete stats Data Error")
-					return needRestart, err
-				}
-			}
-			if r.needApiDel && notDepleted && oldInbound.NodeID == nil {
-				rt, rterr := inboundSvc.runtimeFor(oldInbound)
-				if rterr != nil {
-					needRestart = true
-				} else if err1 := rt.RemoveUser(context.Background(), oldInbound, email); err1 != nil {
-					if !strings.Contains(err1.Error(), fmt.Sprintf("User %s not found.", email)) {
-						needRestart = true
-					}
-				}
-			}
-		}
-		if oldInbound.NodeID != nil && len(email) > 0 {
-			rt, push, dirty, perr := inboundSvc.nodePushPlan(oldInbound)
-			if perr != nil {
-				return needRestart, perr
-			}
-			if dirty {
-				markDirty = true
-			}
-			if push {
-				if err1 := rt.DeleteUser(context.Background(), oldInbound, email); err1 != nil {
-					logger.Warning("Error in deleting client on", rt.Name(), ":", err1)
-					markDirty = true
-				}
-			}
-		}
-	}
-
-	if err := db.Save(oldInbound).Error; err != nil {
-		return needRestart, err
-	}
-	finalClients, gcErr := inboundSvc.GetClients(oldInbound)
-	if gcErr != nil {
-		return needRestart, gcErr
-	}
-	if err := s.SyncInbound(db, inboundId, finalClients); err != nil {
-		return needRestart, err
-	}
-	if markDirty && oldInbound.NodeID != nil {
-		if dErr := (&NodeService{}).MarkNodeDirty(*oldInbound.NodeID); dErr != nil {
-			logger.Warning("mark node dirty failed:", dErr)
-		}
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) DetachByEmailMany(inboundSvc *InboundService, email string, inboundIds []int) (bool, error) {
-	if email == "" {
-		return false, common.NewError("client email is required")
-	}
-	rec, err := s.GetRecordByEmail(nil, email)
-	if err != nil {
-		return false, err
-	}
-	return s.Detach(inboundSvc, rec.Id, inboundIds)
-}
-
-func (s *ClientService) DeleteByEmail(inboundSvc *InboundService, email string, keepTraffic bool) (bool, error) {
-	if email == "" {
-		return false, common.NewError("client email is required")
-	}
-	rec, err := s.GetRecordByEmail(nil, email)
-	if err == nil {
-		return s.Delete(inboundSvc, rec.Id, keepTraffic)
-	}
-	if !errors.Is(err, gorm.ErrRecordNotFound) {
-		return false, err
-	}
-	inboundIds, idsErr := s.findInboundIdsByClientEmail(email)
-	if idsErr != nil {
-		return false, idsErr
-	}
-	if len(inboundIds) == 0 {
-		return false, common.NewError(fmt.Sprintf("client %q not found in any inbound or client record", email))
-	}
-	needRestart := false
-	for _, ibId := range inboundIds {
-		nr, delErr := s.DelInboundClientByEmail(inboundSvc, ibId, email, false)
-		if delErr != nil {
-			if errors.Is(delErr, ErrClientNotInInbound) {
-				continue
-			}
-			return needRestart, delErr
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-	if !keepTraffic {
-		db := database.GetDB()
-		if err := db.Where("email = ?", email).Delete(&xray.ClientTraffic{}).Error; err != nil {
-			return needRestart, err
-		}
-		if err := db.Where("client_email = ?", email).Delete(&model.InboundClientIps{}).Error; err != nil {
-			return needRestart, err
-		}
-	}
-	return needRestart, nil
-}
-
-// findInboundIdsByClientEmail returns every inbound whose settings.clients[]
-// JSON contains an entry with the given email. Driver-portable (no JSON
-// operators) by parsing in Go — fine for the rare fallback path.
-func (s *ClientService) findInboundIdsByClientEmail(email string) ([]int, error) {
-	var inbounds []model.Inbound
-	if err := database.GetDB().
-		Select("id, settings").
-		Where("settings LIKE ?", "%"+email+"%").
-		Find(&inbounds).Error; err != nil {
-		return nil, err
-	}
-	out := make([]int, 0, len(inbounds))
-	for _, ib := range inbounds {
-		var settings map[string]any
-		if err := json.Unmarshal([]byte(ib.Settings), &settings); err != nil {
-			continue
-		}
-		clients, ok := settings["clients"].([]any)
-		if !ok {
-			continue
-		}
-		for _, c := range clients {
-			cm, ok := c.(map[string]any)
-			if !ok {
-				continue
-			}
-			if cEmail, _ := cm["email"].(string); cEmail == email {
-				out = append(out, ib.Id)
-				break
-			}
-		}
-	}
-	return out, nil
-}
-
-func (s *ClientService) UpdateByEmail(inboundSvc *InboundService, email string, updated model.Client, inboundFilter ...int) (bool, error) {
-	if email == "" {
-		return false, common.NewError("client email is required")
-	}
-	rec, err := s.GetRecordByEmail(nil, email)
-	if err != nil {
-		return false, err
-	}
-	return s.Update(inboundSvc, rec.Id, updated, inboundFilter...)
-}
-
-func (s *ClientService) ResetTrafficByEmail(inboundSvc *InboundService, email string) (bool, error) {
-	if email == "" {
-		return false, common.NewError("client email is required")
-	}
-	rec, err := s.GetRecordByEmail(nil, email)
-	if err != nil {
-		return false, err
-	}
-	inboundIds, err := s.GetInboundIdsForRecord(rec.Id)
-	if err != nil {
-		return false, err
-	}
-
-	needRestart := false
-	if !rec.Enable {
-		updated := rec.ToClient()
-		updated.Enable = true
-		nr, uErr := s.Update(inboundSvc, rec.Id, *updated)
-		if uErr != nil {
-			logger.Warning("Failed to auto-enable client during traffic reset:", uErr)
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-
-	if len(inboundIds) == 0 {
-		if rErr := inboundSvc.ResetClientTrafficByEmail(email); rErr != nil {
-			return false, rErr
-		}
-		return needRestart, nil
-	}
-
-	for _, ibId := range inboundIds {
-		nr, rErr := inboundSvc.ResetClientTraffic(ibId, email)
-		if rErr != nil {
-			return needRestart, rErr
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-	return needRestart, nil
-}
-
-// ClientSlim is the row-shape used by the clients page. It drops fields the
-// table never reads (UUID, password, auth, flow, security, reverse, tgId)
-// so the list payload stays compact even when the panel manages thousands
-// of clients. Modals that need the full record still call /get/:email.
-type ClientSlim struct {
-	Email      string              `json:"email"`
-	SubID      string              `json:"subId"`
-	Enable     bool                `json:"enable"`
-	TotalGB    int64               `json:"totalGB"`
-	ExpiryTime int64               `json:"expiryTime"`
-	LimitIP    int                 `json:"limitIp"`
-	Reset      int                 `json:"reset"`
-	Group      string              `json:"group,omitempty"`
-	Comment    string              `json:"comment,omitempty"`
-	InboundIds []int               `json:"inboundIds"`
-	Traffic    *xray.ClientTraffic `json:"traffic,omitempty"`
-	CreatedAt  int64               `json:"createdAt"`
-	UpdatedAt  int64               `json:"updatedAt"`
-}
-
-// ClientPageParams are the query params accepted by /panel/api/clients/list/paged.
-// All fields are optional — the empty value means "no filter" / defaults.
-//
-// Filter / Protocol / Inbound accept either a single value or a comma-separated
-// list; matching is OR within a field and AND across fields. The numeric range
-// fields treat 0 as "unset" on the lower bound and 0 (or negative) as
-// "unbounded" on the upper bound.
-type ClientPageParams struct {
-	Page     int    `form:"page"`
-	PageSize int    `form:"pageSize"`
-	Search   string `form:"search"`
-	Filter   string `form:"filter"`
-	Protocol string `form:"protocol"`
-	Inbound  string `form:"inbound"`
-	Sort     string `form:"sort"`
-	Order    string `form:"order"`
-
-	ExpiryFrom int64  `form:"expiryFrom"`
-	ExpiryTo   int64  `form:"expiryTo"`
-	UsageFrom  int64  `form:"usageFrom"`
-	UsageTo    int64  `form:"usageTo"`
-	AutoRenew  string `form:"autoRenew"`
-	HasTgID    string `form:"hasTgId"`
-	HasComment string `form:"hasComment"`
-	Group      string `form:"group"`
-}
-
-// ClientPageResponse is the shape returned by ListPaged. `Total` is the
-// row count in the DB; `Filtered` is the count after Search/Filter/Protocol
-// were applied, before pagination. The page contains at most PageSize items.
-// Summary is computed across the full DB row set so dashboard counters
-// on the clients page stay stable as the user paginates/filters.
-type ClientPageResponse struct {
-	Items    []ClientSlim   `json:"items"`
-	Total    int            `json:"total"`
-	Filtered int            `json:"filtered"`
-	Page     int            `json:"page"`
-	PageSize int            `json:"pageSize"`
-	Summary  ClientsSummary `json:"summary"`
-	Groups   []string       `json:"groups"`
-}
-
-// ClientsSummary collects per-bucket counts plus the matching email lists so
-// the clients page can render the dashboard stat cards and their hover
-// popovers without shipping the full client array.
-type ClientsSummary struct {
-	Total    int      `json:"total"`
-	Active   int      `json:"active"`
-	Online   []string `json:"online"`
-	Depleted []string `json:"depleted"`
-	Expiring []string `json:"expiring"`
-	Deactive []string `json:"deactive"`
-}
-
-const (
-	clientPageDefaultSize = 25
-	clientPageMaxSize     = 200
-)
-
-// ListPaged loads every client (with traffic + attachments) into memory,
-// applies the requested filter / search / protocol predicates, sorts, and
-// returns the requested page along with total and filtered counts. The DB
-// query itself is unchanged from List(); the win is that the response
-// only carries 25-ish slim rows over the wire instead of all 2000 full
-// records, which on real panels was the dominant cost.
-func (s *ClientService) ListPaged(inboundSvc *InboundService, settingSvc *SettingService, params ClientPageParams) (*ClientPageResponse, error) {
-	all, err := s.List()
-	if err != nil {
-		return nil, err
-	}
-	total := len(all)
-
-	pageSize := params.PageSize
-	if pageSize <= 0 {
-		pageSize = clientPageDefaultSize
-	}
-	if pageSize > clientPageMaxSize {
-		pageSize = clientPageMaxSize
-	}
-	page := params.Page
-	if page <= 0 {
-		page = 1
-	}
-
-	protocols := parseCSVStrings(params.Protocol)
-	inboundIDs := parseCSVInts(params.Inbound)
-	buckets := parseCSVStrings(params.Filter)
-
-	var protocolByInbound map[int]string
-	if len(protocols) > 0 {
-		inbounds, err := inboundSvc.GetAllInbounds()
-		if err == nil {
-			protocolByInbound = make(map[int]string, len(inbounds))
-			for _, ib := range inbounds {
-				protocolByInbound[ib.Id] = string(ib.Protocol)
-			}
-		}
-	}
-
-	onlines := inboundSvc.GetOnlineClients()
-	onlineSet := make(map[string]struct{}, len(onlines))
-	for _, e := range onlines {
-		onlineSet[e] = struct{}{}
-	}
-
-	var expireDiffMs, trafficDiffBytes int64
-	if settingSvc != nil {
-		if v, err := settingSvc.GetExpireDiff(); err == nil {
-			expireDiffMs = int64(v) * 86400000
-		}
-		if v, err := settingSvc.GetTrafficDiff(); err == nil {
-			trafficDiffBytes = int64(v) * 1073741824
-		}
-	}
-
-	nowMs := time.Now().UnixMilli()
-	summary := buildClientsSummary(all, onlineSet, nowMs, expireDiffMs, trafficDiffBytes)
-
-	needle := strings.ToLower(strings.TrimSpace(params.Search))
-
-	filtered := make([]ClientWithAttachments, 0, len(all))
-	for _, c := range all {
-		if needle != "" && !clientMatchesSearch(c, needle) {
-			continue
-		}
-		if len(protocols) > 0 && !clientMatchesAnyProtocol(c, protocols, protocolByInbound) {
-			continue
-		}
-		if len(inboundIDs) > 0 && !clientMatchesAnyInbound(c, inboundIDs) {
-			continue
-		}
-		if len(buckets) > 0 && !clientMatchesAnyBucket(c, buckets, onlineSet, nowMs, expireDiffMs, trafficDiffBytes) {
-			continue
-		}
-		if !clientMatchesExpiryRange(c, params.ExpiryFrom, params.ExpiryTo) {
-			continue
-		}
-		if !clientMatchesUsageRange(c, params.UsageFrom, params.UsageTo) {
-			continue
-		}
-		if !clientMatchesAutoRenew(c, params.AutoRenew) {
-			continue
-		}
-		if !clientMatchesHasTgID(c, params.HasTgID) {
-			continue
-		}
-		if !clientMatchesHasComment(c, params.HasComment) {
-			continue
-		}
-		if !clientMatchesAnyGroup(c, params.Group) {
-			continue
-		}
-		filtered = append(filtered, c)
-	}
-
-	sortClients(filtered, params.Sort, params.Order)
-
-	filteredCount := len(filtered)
-	start := (page - 1) * pageSize
-	end := start + pageSize
-	if start > filteredCount {
-		start = filteredCount
-	}
-	if end > filteredCount {
-		end = filteredCount
-	}
-	pageRows := filtered[start:end]
-
-	items := make([]ClientSlim, 0, len(pageRows))
-	for _, c := range pageRows {
-		items = append(items, toClientSlim(c))
-	}
-
-	groupRows, gErr := s.ListGroups()
-	if gErr != nil {
-		return nil, gErr
-	}
-	groups := make([]string, 0, len(groupRows))
-	for _, g := range groupRows {
-		groups = append(groups, g.Name)
-	}
-
-	return &ClientPageResponse{
-		Items:    items,
-		Total:    total,
-		Filtered: filteredCount,
-		Page:     page,
-		PageSize: pageSize,
-		Summary:  summary,
-		Groups:   groups,
-	}, nil
-}
-
-type GroupSummary struct {
-	Name        string `json:"name"`
-	ClientCount int    `json:"clientCount"`
-	TrafficUsed int64  `json:"trafficUsed"`
-}
-
-func (s *ClientService) ListGroups() ([]GroupSummary, error) {
-	db := database.GetDB()
-	// email is unique in both clients and client_traffics, so the LEFT JOIN
-	// never double-counts a client's traffic.
-	var derived []GroupSummary
-	if err := db.Table("clients AS c").
-		Select("c.group_name AS name, COUNT(*) AS client_count, COALESCE(SUM(ct.up + ct.down), 0) AS traffic_used").
-		Joins("LEFT JOIN client_traffics ct ON ct.email = c.email").
-		Where("c.group_name <> ''").
-		Group("c.group_name").
-		Scan(&derived).Error; err != nil {
-		return nil, err
-	}
-	var stored []model.ClientGroup
-	if err := db.Find(&stored).Error; err != nil {
-		return nil, err
-	}
-	type groupAgg struct {
-		count   int
-		traffic int64
-	}
-	merged := make(map[string]groupAgg, len(derived)+len(stored))
-	for _, g := range stored {
-		merged[g.Name] = groupAgg{}
-	}
-	for _, g := range derived {
-		merged[g.Name] = groupAgg{count: g.ClientCount, traffic: g.TrafficUsed}
-	}
-	out := make([]GroupSummary, 0, len(merged))
-	for name, agg := range merged {
-		out = append(out, GroupSummary{Name: name, ClientCount: agg.count, TrafficUsed: agg.traffic})
-	}
-	sort.Slice(out, func(i, j int) bool {
-		return strings.ToLower(out[i].Name) < strings.ToLower(out[j].Name)
-	})
-	return out, nil
-}
-
-func (s *ClientService) EmailsByGroup(name string) ([]string, error) {
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return []string{}, nil
-	}
-	db := database.GetDB()
-	var emails []string
-	if err := db.Model(&model.ClientRecord{}).
-		Where("group_name = ?", name).
-		Order("email ASC").
-		Pluck("email", &emails).Error; err != nil {
-		return nil, err
-	}
-	if emails == nil {
-		emails = []string{}
-	}
-	return emails, nil
-}
-
-func (s *ClientService) BulkResetTraffic(inboundSvc *InboundService, emails []string) (int, error) {
-	if len(emails) == 0 {
-		return 0, nil
-	}
-	seen := map[string]struct{}{}
-	cleanEmails := make([]string, 0, len(emails))
-	for _, e := range emails {
-		e = strings.TrimSpace(e)
-		if e == "" {
-			continue
-		}
-		if _, ok := seen[e]; ok {
-			continue
-		}
-		seen[e] = struct{}{}
-		cleanEmails = append(cleanEmails, e)
-	}
-	if len(cleanEmails) == 0 {
-		return 0, nil
-	}
-
-	for _, e := range cleanEmails {
-		rec, err := s.GetRecordByEmail(nil, e)
-		if err == nil && !rec.Enable {
-			updated := rec.ToClient()
-			updated.Enable = true
-			s.Update(inboundSvc, rec.Id, *updated)
-		}
-	}
-
-	affected := 0
-	err := submitTrafficWrite(func() error {
-		db := database.GetDB()
-		return db.Transaction(func(tx *gorm.DB) error {
-			for _, batch := range chunkStrings(cleanEmails, sqlInChunk) {
-				res := tx.Model(xray.ClientTraffic{}).
-					Where("email IN ?", batch).
-					Updates(map[string]any{"enable": true, "up": 0, "down": 0})
-				if res.Error != nil {
-					return res.Error
-				}
-				affected += int(res.RowsAffected)
-			}
-			return nil
-		})
-	})
-	if err != nil {
-		return 0, err
-	}
-	return affected, nil
-}
-
-func (s *ClientService) CreateGroup(name string) error {
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return common.NewError("group name is required")
-	}
-	db := database.GetDB()
-	var count int64
-	if err := db.Model(&model.ClientGroup{}).Where("name = ?", name).Count(&count).Error; err != nil {
-		return err
-	}
-	if count > 0 {
-		return common.NewError("group already exists")
-	}
-	return db.Create(&model.ClientGroup{Name: name}).Error
-}
-
-func (s *ClientService) RenameGroup(oldName, newName string) (int, error) {
-	oldName = strings.TrimSpace(oldName)
-	newName = strings.TrimSpace(newName)
-	if oldName == "" {
-		return 0, common.NewError("old group name is required")
-	}
-	if newName == "" {
-		return 0, common.NewError("new group name is required")
-	}
-	if oldName == newName {
-		return 0, nil
-	}
-	return s.replaceGroupValue(oldName, newName)
-}
-
-func (s *ClientService) DeleteGroup(name string) (int, error) {
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return 0, common.NewError("group name is required")
-	}
-	return s.replaceGroupValue(name, "")
-}
-
-func (s *ClientService) RemoveFromGroup(emails []string) (int, error) {
-	return s.AddToGroup(emails, "")
-}
-
-func (s *ClientService) AddToGroup(emails []string, group string) (int, error) {
-	group = strings.TrimSpace(group)
-	if len(emails) == 0 {
-		return 0, nil
-	}
-	db := database.GetDB()
-
-	if group != "" {
-		var exists int64
-		if err := db.Model(&model.ClientGroup{}).Where("name = ?", group).Count(&exists).Error; err != nil {
-			return 0, err
-		}
-		if exists == 0 {
-			var derived int64
-			if err := db.Model(&model.ClientRecord{}).Where("group_name = ?", group).Count(&derived).Error; err != nil {
-				return 0, err
-			}
-			if derived == 0 {
-				if err := db.Create(&model.ClientGroup{Name: group}).Error; err != nil {
-					return 0, err
-				}
-			}
-		}
-	}
-
-	var records []model.ClientRecord
-	for _, batch := range chunkStrings(emails, sqlInChunk) {
-		var rows []model.ClientRecord
-		if err := db.Where("email IN ?", batch).Find(&rows).Error; err != nil {
-			return 0, err
-		}
-		records = append(records, rows...)
-	}
-	if len(records) == 0 {
-		return 0, nil
-	}
-	affectedEmails := make([]string, 0, len(records))
-	for _, r := range records {
-		affectedEmails = append(affectedEmails, r.Email)
-	}
-
-	tx := db.Begin()
-	for _, batch := range chunkStrings(affectedEmails, sqlInChunk) {
-		if err := tx.Model(&model.ClientRecord{}).
-			Where("email IN ?", batch).
-			UpdateColumn("group_name", group).Error; err != nil {
-			tx.Rollback()
-			return 0, err
-		}
-	}
-
-	var inboundIDs []int
-	inboundIDSeen := make(map[int]struct{})
-	for _, batch := range chunkStrings(affectedEmails, sqlInChunk) {
-		var ids []int
-		if err := tx.Table("client_inbounds").
-			Joins("JOIN clients ON clients.id = client_inbounds.client_id").
-			Where("clients.email IN ?", batch).
-			Distinct("client_inbounds.inbound_id").
-			Pluck("inbound_id", &ids).Error; err != nil {
-			tx.Rollback()
-			return 0, err
-		}
-		for _, id := range ids {
-			if _, ok := inboundIDSeen[id]; !ok {
-				inboundIDSeen[id] = struct{}{}
-				inboundIDs = append(inboundIDs, id)
-			}
-		}
-	}
-
-	emailSet := make(map[string]struct{}, len(affectedEmails))
-	for _, e := range affectedEmails {
-		emailSet[e] = struct{}{}
-	}
-
-	for _, ibID := range inboundIDs {
-		var ib model.Inbound
-		if err := tx.First(&ib, ibID).Error; err != nil {
-			tx.Rollback()
-			return 0, err
-		}
-		var settings map[string]any
-		if err := json.Unmarshal([]byte(ib.Settings), &settings); err != nil {
-			continue
-		}
-		clients, ok := settings["clients"].([]any)
-		if !ok {
-			continue
-		}
-		modified := false
-		for i := range clients {
-			cm, ok := clients[i].(map[string]any)
-			if !ok {
-				continue
-			}
-			email, _ := cm["email"].(string)
-			if _, hit := emailSet[email]; !hit {
-				continue
-			}
-			if group == "" {
-				delete(cm, "group")
-			} else {
-				cm["group"] = group
-			}
-			clients[i] = cm
-			modified = true
-		}
-		if modified {
-			settings["clients"] = clients
-			newSettings, err := json.Marshal(settings)
-			if err != nil {
-				continue
-			}
-			ib.Settings = string(newSettings)
-			if err := tx.Save(&ib).Error; err != nil {
-				tx.Rollback()
-				return 0, err
-			}
-		}
-	}
-
-	if err := tx.Commit().Error; err != nil {
-		return 0, err
-	}
-	return len(records), nil
-}
-
-func (s *ClientService) replaceGroupValue(oldName, newName string) (int, error) {
-	db := database.GetDB()
-	if newName == "" {
-		if err := db.Where("name = ?", oldName).Delete(&model.ClientGroup{}).Error; err != nil {
-			return 0, err
-		}
-	} else {
-		if err := db.Model(&model.ClientGroup{}).Where("name = ?", oldName).Update("name", newName).Error; err != nil {
-			return 0, err
-		}
-	}
-	var records []model.ClientRecord
-	if err := db.Where("group_name = ?", oldName).Find(&records).Error; err != nil {
-		return 0, err
-	}
-	if len(records) == 0 {
-		return 0, nil
-	}
-	affectedEmails := make([]string, 0, len(records))
-	for _, r := range records {
-		affectedEmails = append(affectedEmails, r.Email)
-	}
-
-	tx := db.Begin()
-	if err := tx.Model(&model.ClientRecord{}).
-		Where("group_name = ?", oldName).
-		UpdateColumn("group_name", newName).Error; err != nil {
-		tx.Rollback()
-		return 0, err
-	}
-
-	var inboundIDs []int
-	inboundIDSeen := make(map[int]struct{})
-	for _, batch := range chunkStrings(affectedEmails, sqlInChunk) {
-		var ids []int
-		if err := tx.Table("client_inbounds").
-			Joins("JOIN clients ON clients.id = client_inbounds.client_id").
-			Where("clients.email IN ?", batch).
-			Distinct("client_inbounds.inbound_id").
-			Pluck("inbound_id", &ids).Error; err != nil {
-			tx.Rollback()
-			return 0, err
-		}
-		for _, id := range ids {
-			if _, ok := inboundIDSeen[id]; !ok {
-				inboundIDSeen[id] = struct{}{}
-				inboundIDs = append(inboundIDs, id)
-			}
-		}
-	}
-
-	for _, ibID := range inboundIDs {
-		var ib model.Inbound
-		if err := tx.First(&ib, ibID).Error; err != nil {
-			tx.Rollback()
-			return 0, err
-		}
-		var settings map[string]any
-		if err := json.Unmarshal([]byte(ib.Settings), &settings); err != nil {
-			continue
-		}
-		clients, ok := settings["clients"].([]any)
-		if !ok {
-			continue
-		}
-		modified := false
-		for i := range clients {
-			cm, ok := clients[i].(map[string]any)
-			if !ok {
-				continue
-			}
-			if g, ok := cm["group"].(string); ok && g == oldName {
-				if newName == "" {
-					delete(cm, "group")
-				} else {
-					cm["group"] = newName
-				}
-				clients[i] = cm
-				modified = true
-			}
-		}
-		if modified {
-			settings["clients"] = clients
-			newSettings, err := json.Marshal(settings)
-			if err != nil {
-				continue
-			}
-			ib.Settings = string(newSettings)
-			if err := tx.Save(&ib).Error; err != nil {
-				tx.Rollback()
-				return 0, err
-			}
-		}
-	}
-
-	if err := tx.Commit().Error; err != nil {
-		return 0, err
-	}
-	return len(records), nil
-}
-
-func buildClientsSummary(all []ClientWithAttachments, onlineSet map[string]struct{}, nowMs, expireDiffMs, trafficDiffBytes int64) ClientsSummary {
-	s := ClientsSummary{
-		Total:    len(all),
-		Online:   []string{},
-		Depleted: []string{},
-		Expiring: []string{},
-		Deactive: []string{},
-	}
-	for _, c := range all {
-		used := int64(0)
-		if c.Traffic != nil {
-			used = c.Traffic.Up + c.Traffic.Down
-		}
-		exhausted := c.TotalGB > 0 && used >= c.TotalGB
-		expired := c.ExpiryTime > 0 && c.ExpiryTime <= nowMs
-		if c.Enable {
-			if _, ok := onlineSet[c.Email]; ok {
-				s.Online = append(s.Online, c.Email)
-			}
-		}
-		if exhausted || expired {
-			s.Depleted = append(s.Depleted, c.Email)
-			continue
-		}
-		if !c.Enable {
-			s.Deactive = append(s.Deactive, c.Email)
-			continue
-		}
-		nearExpiry := c.ExpiryTime > 0 && c.ExpiryTime-nowMs < expireDiffMs
-		nearLimit := c.TotalGB > 0 && c.TotalGB-used < trafficDiffBytes
-		if nearExpiry || nearLimit {
-			s.Expiring = append(s.Expiring, c.Email)
-		} else {
-			s.Active++
-		}
-	}
-	return s
-}
-
-func toClientSlim(c ClientWithAttachments) ClientSlim {
-	return ClientSlim{
-		Email:      c.Email,
-		SubID:      c.SubID,
-		Enable:     c.Enable,
-		TotalGB:    c.TotalGB,
-		ExpiryTime: c.ExpiryTime,
-		LimitIP:    c.LimitIP,
-		Reset:      c.Reset,
-		Group:      c.Group,
-		Comment:    c.Comment,
-		InboundIds: c.InboundIds,
-		Traffic:    c.Traffic,
-		CreatedAt:  c.CreatedAt,
-		UpdatedAt:  c.UpdatedAt,
-	}
-}
-
-func clientMatchesSearch(c ClientWithAttachments, needle string) bool {
-	if needle == "" {
-		return true
-	}
-	candidates := [...]string{c.Email, c.SubID, c.Comment, c.UUID, c.Password, c.Auth}
-	for _, v := range candidates {
-		if v != "" && strings.Contains(strings.ToLower(v), needle) {
-			return true
-		}
-	}
-	return false
-}
-
-// parseCSVStrings splits a comma-separated list, trims/lower-cases each item,
-// and drops blanks. Returns nil when the input has no usable entries — the
-// caller can then skip the predicate entirely.
-func parseCSVStrings(raw string) []string {
-	if raw == "" {
-		return nil
-	}
-	parts := strings.Split(raw, ",")
-	out := make([]string, 0, len(parts))
-	for _, p := range parts {
-		s := strings.ToLower(strings.TrimSpace(p))
-		if s != "" {
-			out = append(out, s)
-		}
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-// parseCSVInts is parseCSVStrings for positive integer IDs; non-numeric or
-// non-positive entries are silently dropped.
-func parseCSVInts(raw string) []int {
-	if raw == "" {
-		return nil
-	}
-	parts := strings.Split(raw, ",")
-	out := make([]int, 0, len(parts))
-	for _, p := range parts {
-		s := strings.TrimSpace(p)
-		if s == "" {
-			continue
-		}
-		if n, err := strconv.Atoi(s); err == nil && n > 0 {
-			out = append(out, n)
-		}
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-func clientMatchesAnyProtocol(c ClientWithAttachments, protocols []string, byInbound map[int]string) bool {
-	for _, id := range c.InboundIds {
-		p := byInbound[id]
-		if p == "" {
-			continue
-		}
-		if slices.Contains(protocols, strings.ToLower(p)) {
-			return true
-		}
-	}
-	return false
-}
-
-func clientMatchesAnyInbound(c ClientWithAttachments, inboundIds []int) bool {
-	for _, id := range c.InboundIds {
-		if slices.Contains(inboundIds, id) {
-			return true
-		}
-	}
-	return false
-}
-
-func clientMatchesAnyBucket(c ClientWithAttachments, buckets []string, onlineSet map[string]struct{}, nowMs, expireDiffMs, trafficDiffBytes int64) bool {
-	for _, b := range buckets {
-		if clientMatchesBucket(c, b, onlineSet, nowMs, expireDiffMs, trafficDiffBytes) {
-			return true
-		}
-	}
-	return false
-}
-
-func clientMatchesExpiryRange(c ClientWithAttachments, fromMs, toMs int64) bool {
-	if fromMs <= 0 && toMs <= 0 {
-		return true
-	}
-	// expiryTime of 0 means "never expires"; treat it as outside any bounded
-	// range so users filtering by date see only clients with concrete expiries.
-	if c.ExpiryTime == 0 {
-		return false
-	}
-	// Negative expiry is the "delayed start" sentinel; same treatment as never.
-	if c.ExpiryTime < 0 {
-		return false
-	}
-	if fromMs > 0 && c.ExpiryTime < fromMs {
-		return false
-	}
-	if toMs > 0 && c.ExpiryTime > toMs {
-		return false
-	}
-	return true
-}
-
-func clientMatchesUsageRange(c ClientWithAttachments, fromBytes, toBytes int64) bool {
-	if fromBytes <= 0 && toBytes <= 0 {
-		return true
-	}
-	used := int64(0)
-	if c.Traffic != nil {
-		used = c.Traffic.Up + c.Traffic.Down
-	}
-	if fromBytes > 0 && used < fromBytes {
-		return false
-	}
-	if toBytes > 0 && used > toBytes {
-		return false
-	}
-	return true
-}
-
-func clientMatchesAutoRenew(c ClientWithAttachments, mode string) bool {
-	switch strings.ToLower(strings.TrimSpace(mode)) {
-	case "on":
-		return c.Reset > 0
-	case "off":
-		return c.Reset <= 0
-	}
-	return true
-}
-
-func clientMatchesHasTgID(c ClientWithAttachments, mode string) bool {
-	switch strings.ToLower(strings.TrimSpace(mode)) {
-	case "yes":
-		return c.TgID != 0
-	case "no":
-		return c.TgID == 0
-	}
-	return true
-}
-
-func clientMatchesHasComment(c ClientWithAttachments, mode string) bool {
-	switch strings.ToLower(strings.TrimSpace(mode)) {
-	case "yes":
-		return strings.TrimSpace(c.Comment) != ""
-	case "no":
-		return strings.TrimSpace(c.Comment) == ""
-	}
-	return true
-}
-
-func clientMatchesAnyGroup(c ClientWithAttachments, csv string) bool {
-	groups := parseCSVStrings(csv)
-	if len(groups) == 0 {
-		return true
-	}
-	current := strings.TrimSpace(c.Group)
-	for _, g := range groups {
-		if g == "" {
-			if current == "" {
-				return true
-			}
-			continue
-		}
-		if strings.EqualFold(g, current) {
-			return true
-		}
-	}
-	return false
-}
-
-func clientMatchesBucket(c ClientWithAttachments, bucket string, onlineSet map[string]struct{}, nowMs, expireDiffMs, trafficDiffBytes int64) bool {
-	if bucket == "" {
-		return true
-	}
-	used := int64(0)
-	if c.Traffic != nil {
-		used = c.Traffic.Up + c.Traffic.Down
-	}
-	exhausted := c.TotalGB > 0 && used >= c.TotalGB
-	expired := c.ExpiryTime > 0 && c.ExpiryTime <= nowMs
-	switch bucket {
-	case "online":
-		if onlineSet == nil {
-			return false
-		}
-		_, ok := onlineSet[c.Email]
-		return ok && c.Enable
-	case "depleted":
-		return exhausted || expired
-	case "deactive":
-		return !c.Enable
-	case "active":
-		return c.Enable && !exhausted && !expired
-	case "expiring":
-		if !c.Enable || exhausted || expired {
-			return false
-		}
-		nearExpiry := c.ExpiryTime > 0 && c.ExpiryTime-nowMs < expireDiffMs
-		nearLimit := c.TotalGB > 0 && c.TotalGB-used < trafficDiffBytes
-		return nearExpiry || nearLimit
-	}
-	return true
-}
-
-func sortClients(rows []ClientWithAttachments, sortKey, order string) {
-	if sortKey == "" {
-		return
-	}
-	desc := order == "descend"
-	less := func(i, j int) bool {
-		a, b := rows[i], rows[j]
-		switch sortKey {
-		case "enable":
-			if a.Enable == b.Enable {
-				return false
-			}
-			return !a.Enable && b.Enable
-		case "email":
-			return strings.ToLower(a.Email) < strings.ToLower(b.Email)
-		case "inboundIds":
-			return len(a.InboundIds) < len(b.InboundIds)
-		case "traffic":
-			ua := int64(0)
-			if a.Traffic != nil {
-				ua = a.Traffic.Up + a.Traffic.Down
-			}
-			ub := int64(0)
-			if b.Traffic != nil {
-				ub = b.Traffic.Up + b.Traffic.Down
-			}
-			return ua < ub
-		case "remaining":
-			ra := int64(1<<62 - 1)
-			if a.TotalGB > 0 {
-				used := int64(0)
-				if a.Traffic != nil {
-					used = a.Traffic.Up + a.Traffic.Down
-				}
-				ra = a.TotalGB - used
-			}
-			rb := int64(1<<62 - 1)
-			if b.TotalGB > 0 {
-				used := int64(0)
-				if b.Traffic != nil {
-					used = b.Traffic.Up + b.Traffic.Down
-				}
-				rb = b.TotalGB - used
-			}
-			return ra < rb
-		case "expiryTime":
-			ea := int64(1<<62 - 1)
-			if a.ExpiryTime > 0 {
-				ea = a.ExpiryTime
-			}
-			eb := int64(1<<62 - 1)
-			if b.ExpiryTime > 0 {
-				eb = b.ExpiryTime
-			}
-			return ea < eb
-		case "createdAt":
-			if a.CreatedAt == b.CreatedAt {
-				return a.Id < b.Id
-			}
-			return a.CreatedAt < b.CreatedAt
-		case "updatedAt":
-			if a.UpdatedAt == b.UpdatedAt {
-				return a.Id < b.Id
-			}
-			return a.UpdatedAt < b.UpdatedAt
-		case "lastOnline":
-			la := int64(0)
-			if a.Traffic != nil {
-				la = a.Traffic.LastOnline
-			}
-			lb := int64(0)
-			if b.Traffic != nil {
-				lb = b.Traffic.LastOnline
-			}
-			if la == lb {
-				return a.Id < b.Id
-			}
-			return la < lb
-		}
-		return false
-	}
-	sort.SliceStable(rows, func(i, j int) bool {
-		if desc {
-			return less(j, i)
-		}
-		return less(i, j)
-	})
-}
-
-// BulkAdjustResult is returned by BulkAdjust to report how many clients were
-// successfully updated and which were skipped (typically because the field
-// being adjusted was unlimited for that client) or failed.
-type BulkAdjustResult struct {
-	Adjusted int                `json:"adjusted"`
-	Skipped  []BulkAdjustReport `json:"skipped,omitempty"`
-}
-
-type BulkAdjustReport struct {
-	Email  string `json:"email"`
-	Reason string `json:"reason"`
-}
-
-type bulkAdjustEntry struct {
-	record      *model.ClientRecord
-	applyExpiry bool
-	newExpiry   int64
-	applyTotal  bool
-	newTotal    int64
-}
-
-// BulkAdjust shifts ExpiryTime by addDays (days) and TotalGB by addBytes
-// for every email in the list. Clients whose corresponding field is
-// unlimited (0) are skipped — bulk extend should not accidentally
-// limit an unlimited client. addDays and addBytes may be negative.
-//
-// Like BulkDelete, the work is grouped by inbound so each inbound's
-// settings JSON is parsed and written exactly once regardless of how
-// many target emails it contains.
-func (s *ClientService) BulkAdjust(inboundSvc *InboundService, emails []string, addDays int, addBytes int64) (BulkAdjustResult, bool, error) {
-	result := BulkAdjustResult{}
-	if len(emails) == 0 {
-		return result, false, nil
-	}
-	if addDays == 0 && addBytes == 0 {
-		return result, false, common.NewError("no adjustment specified")
-	}
-
-	addExpiryMs := int64(addDays) * 24 * 60 * 60 * 1000
-
-	seen := map[string]struct{}{}
-	cleanEmails := make([]string, 0, len(emails))
-	for _, e := range emails {
-		e = strings.TrimSpace(e)
-		if e == "" {
-			continue
-		}
-		if _, ok := seen[e]; ok {
-			continue
-		}
-		seen[e] = struct{}{}
-		cleanEmails = append(cleanEmails, e)
-	}
-	if len(cleanEmails) == 0 {
-		return result, false, nil
-	}
-
-	db := database.GetDB()
-
-	var records []model.ClientRecord
-	for _, batch := range chunkStrings(cleanEmails, sqlInChunk) {
-		var rows []model.ClientRecord
-		if err := db.Where("email IN ?", batch).Find(&rows).Error; err != nil {
-			return result, false, err
-		}
-		records = append(records, rows...)
-	}
-	recordsByEmail := make(map[string]*model.ClientRecord, len(records))
-	for i := range records {
-		recordsByEmail[records[i].Email] = &records[i]
-	}
-
-	skippedReasons := map[string]string{}
-	for _, email := range cleanEmails {
-		if _, ok := recordsByEmail[email]; !ok {
-			skippedReasons[email] = "client not found"
-		}
-	}
-
-	plan := map[string]*bulkAdjustEntry{}
-	for email, rec := range recordsByEmail {
-		entry := &bulkAdjustEntry{record: rec}
-		if addDays != 0 {
-			switch {
-			case rec.ExpiryTime == 0:
-				if _, exists := skippedReasons[email]; !exists {
-					skippedReasons[email] = "unlimited expiry"
-				}
-			case rec.ExpiryTime > 0:
-				next := rec.ExpiryTime + addExpiryMs
-				if next <= 0 {
-					if _, exists := skippedReasons[email]; !exists {
-						skippedReasons[email] = "reduction exceeds remaining time"
-					}
-				} else {
-					entry.applyExpiry = true
-					entry.newExpiry = next
-				}
-			default:
-				next := rec.ExpiryTime - addExpiryMs
-				if next >= 0 {
-					if _, exists := skippedReasons[email]; !exists {
-						skippedReasons[email] = "reduction exceeds delay window"
-					}
-				} else {
-					entry.applyExpiry = true
-					entry.newExpiry = next
-				}
-			}
-		}
-		if addBytes != 0 {
-			if rec.TotalGB == 0 {
-				if _, exists := skippedReasons[email]; !exists {
-					skippedReasons[email] = "unlimited traffic"
-				}
-			} else {
-				next := max(rec.TotalGB+addBytes, 0)
-				entry.applyTotal = true
-				entry.newTotal = next
-			}
-		}
-		if entry.applyExpiry || entry.applyTotal {
-			plan[email] = entry
-		}
-	}
-
-	if len(plan) == 0 {
-		for email, reason := range skippedReasons {
-			result.Skipped = append(result.Skipped, BulkAdjustReport{Email: email, Reason: reason})
-		}
-		return result, false, nil
-	}
-
-	plannedIds := make([]int, 0, len(plan))
-	recordIdToEmail := make(map[int]string, len(plan))
-	for email, entry := range plan {
-		plannedIds = append(plannedIds, entry.record.Id)
-		recordIdToEmail[entry.record.Id] = email
-	}
-
-	var mappings []model.ClientInbound
-	for _, batch := range chunkInts(plannedIds, sqlInChunk) {
-		var rows []model.ClientInbound
-		if err := db.Where("client_id IN ?", batch).Find(&rows).Error; err != nil {
-			return result, false, err
-		}
-		mappings = append(mappings, rows...)
-	}
-	emailsByInbound := map[int][]string{}
-	for _, m := range mappings {
-		email, ok := recordIdToEmail[m.ClientId]
-		if !ok {
-			continue
-		}
-		emailsByInbound[m.InboundId] = append(emailsByInbound[m.InboundId], email)
-	}
-
-	needRestart := false
-	for inboundId, ibEmails := range emailsByInbound {
-		ibRes := s.bulkAdjustInboundClients(inboundSvc, inboundId, ibEmails, plan)
-		if ibRes.needRestart {
-			needRestart = true
-		}
-		for email, reason := range ibRes.perEmailSkipped {
-			if _, already := skippedReasons[email]; !already {
-				skippedReasons[email] = reason
-			}
-		}
-	}
-
-	for email, entry := range plan {
-		if _, skipped := skippedReasons[email]; skipped {
-			continue
-		}
-		updates := map[string]any{}
-		if entry.applyExpiry {
-			updates["expiry_time"] = entry.newExpiry
-		}
-		if entry.applyTotal {
-			updates["total"] = entry.newTotal
-		}
-		if len(updates) == 0 {
-			continue
-		}
-		if err := db.Model(xray.ClientTraffic{}).Where("email = ?", email).Updates(updates).Error; err != nil {
-			if _, already := skippedReasons[email]; !already {
-				skippedReasons[email] = err.Error()
-			}
-			continue
-		}
-		result.Adjusted++
-	}
-
-	for email, reason := range skippedReasons {
-		result.Skipped = append(result.Skipped, BulkAdjustReport{Email: email, Reason: reason})
-	}
-	return result, needRestart, nil
-}
-
-type bulkInboundAdjustResult struct {
-	perEmailSkipped map[string]string
-	needRestart     bool
-}
-
-// bulkAdjustInboundClients applies expiry/total deltas to multiple clients
-// inside a single inbound's settings JSON. The xray runtime is updated
-// only for remote-node inbounds; local nodes do not need a notification
-// because the AddUser payload does not include totalGB/expiryTime —
-// changing those fields is identity-preserving and the panel's traffic
-// enforcement loop picks up the new limits from ClientTraffic directly.
-func (s *ClientService) bulkAdjustInboundClients(
-	inboundSvc *InboundService,
-	inboundId int,
-	emails []string,
-	plan map[string]*bulkAdjustEntry,
-) bulkInboundAdjustResult {
-	res := bulkInboundAdjustResult{perEmailSkipped: map[string]string{}}
-
-	defer lockInbound(inboundId).Unlock()
-
-	oldInbound, err := inboundSvc.GetInbound(inboundId)
-	if err != nil {
-		logger.Error("Load Old Data Error")
-		for _, e := range emails {
-			res.perEmailSkipped[e] = err.Error()
-		}
-		return res
-	}
-
-	var settings map[string]any
-	if err := json.Unmarshal([]byte(oldInbound.Settings), &settings); err != nil {
-		for _, e := range emails {
-			res.perEmailSkipped[e] = err.Error()
-		}
-		return res
-	}
-
-	// Match by email — the client's stable identity (see Delete). Credentials
-	// can drift from the inbound JSON, so they are never used for matching.
-	wantedEmails := make(map[string]struct{}, len(emails))
-	for _, email := range emails {
-		if plan[email] == nil {
-			res.perEmailSkipped[email] = "client not found"
-			continue
-		}
-		wantedEmails[email] = struct{}{}
-	}
-
-	interfaceClients, _ := settings["clients"].([]any)
-	foundEmails := map[string]bool{}
-	nowMs := time.Now().Unix() * 1000
-	for i, client := range interfaceClients {
-		c, ok := client.(map[string]any)
-		if !ok {
-			continue
-		}
-		targetEmail, _ := c["email"].(string)
-		if _, want := wantedEmails[targetEmail]; !want || targetEmail == "" {
-			continue
-		}
-		entry := plan[targetEmail]
-		if entry.applyExpiry {
-			c["expiryTime"] = entry.newExpiry
-		}
-		if entry.applyTotal {
-			c["totalGB"] = entry.newTotal
-		}
-		c["updated_at"] = nowMs
-		interfaceClients[i] = c
-		foundEmails[targetEmail] = true
-	}
-
-	for email := range wantedEmails {
-		if !foundEmails[email] {
-			res.perEmailSkipped[email] = "Client Not Found In Inbound"
-		}
-	}
-
-	if len(foundEmails) == 0 {
-		return res
-	}
-
-	settings["clients"] = interfaceClients
-	newSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		for email := range foundEmails {
-			res.perEmailSkipped[email] = err.Error()
-		}
-		return res
-	}
-	oldInbound.Settings = string(newSettings)
-
-	markDirty := false
-	if oldInbound.NodeID != nil {
-		rt, push, dirty, perr := inboundSvc.nodePushPlan(oldInbound)
-		if perr != nil {
-			for email := range foundEmails {
-				res.perEmailSkipped[email] = perr.Error()
-				delete(foundEmails, email)
-			}
-		} else {
-			if dirty {
-				markDirty = true
-			}
-			if push {
-				for email := range foundEmails {
-					entry := plan[email]
-					updated := *entry.record.ToClient()
-					if entry.applyExpiry {
-						updated.ExpiryTime = entry.newExpiry
-					}
-					if entry.applyTotal {
-						updated.TotalGB = entry.newTotal
-					}
-					updated.UpdatedAt = nowMs
-					if err1 := rt.UpdateUser(context.Background(), oldInbound, email, updated); err1 != nil {
-						logger.Warning("Error in updating client on", rt.Name(), ":", err1)
-						markDirty = true
-					}
-				}
-			}
-		}
-	}
-
-	db := database.GetDB()
-	txErr := db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Save(oldInbound).Error; err != nil {
-			return err
-		}
-		finalClients, gcErr := inboundSvc.GetClients(oldInbound)
-		if gcErr != nil {
-			return gcErr
-		}
-		return s.SyncInbound(tx, inboundId, finalClients)
-	})
-	if txErr != nil {
-		for email := range foundEmails {
-			if _, skip := res.perEmailSkipped[email]; !skip {
-				res.perEmailSkipped[email] = txErr.Error()
-			}
-		}
-	} else if markDirty && oldInbound.NodeID != nil {
-		if dErr := (&NodeService{}).MarkNodeDirty(*oldInbound.NodeID); dErr != nil {
-			logger.Warning("mark node dirty failed:", dErr)
-		}
-	}
-
-	return res
-}
-
-// BulkDeleteResult mirrors BulkAdjustResult: total deleted plus per-email
-// skip reasons when an email could not be processed.
-type BulkDeleteResult struct {
-	Deleted int                `json:"deleted"`
-	Skipped []BulkDeleteReport `json:"skipped,omitempty"`
-}
-
-type BulkDeleteReport struct {
-	Email  string `json:"email"`
-	Reason string `json:"reason"`
-}
-
 const sqlInChunk = 400
-
-// BulkDelete removes every client in the list in one optimized pass.
-// Instead of running the full single-delete pipeline N times (which would
-// re-read, re-parse, and re-write each inbound's settings JSON for every
-// email), it groups emails by inbound and performs a single
-// read-modify-write per inbound. Per-row DB cleanups are also batched with
-// IN-clause queries at the end. Errors on a particular email are recorded
-// in the Skipped list and processing continues for the rest.
-func (s *ClientService) BulkDelete(inboundSvc *InboundService, emails []string, keepTraffic bool) (BulkDeleteResult, bool, error) {
-	result := BulkDeleteResult{}
-
-	seen := map[string]struct{}{}
-	cleanEmails := make([]string, 0, len(emails))
-	for _, e := range emails {
-		e = strings.TrimSpace(e)
-		if e == "" {
-			continue
-		}
-		if _, ok := seen[e]; ok {
-			continue
-		}
-		seen[e] = struct{}{}
-		cleanEmails = append(cleanEmails, e)
-	}
-	if len(cleanEmails) == 0 {
-		return result, false, nil
-	}
-
-	db := database.GetDB()
-
-	var records []model.ClientRecord
-	for _, batch := range chunkStrings(cleanEmails, sqlInChunk) {
-		var rows []model.ClientRecord
-		if err := db.Where("email IN ?", batch).Find(&rows).Error; err != nil {
-			return result, false, err
-		}
-		records = append(records, rows...)
-	}
-	recordsByEmail := make(map[string]*model.ClientRecord, len(records))
-	tombstoneEmails := make([]string, 0, len(records))
-	for i := range records {
-		recordsByEmail[records[i].Email] = &records[i]
-		tombstoneEmails = append(tombstoneEmails, records[i].Email)
-	}
-	tombstoneClientEmails(tombstoneEmails)
-
-	skippedReasons := map[string]string{}
-	for _, email := range cleanEmails {
-		if _, ok := recordsByEmail[email]; !ok {
-			skippedReasons[email] = "client not found"
-		}
-	}
-
-	clientIds := make([]int, 0, len(recordsByEmail))
-	recordIdToEmail := make(map[int]string, len(recordsByEmail))
-	for _, r := range recordsByEmail {
-		clientIds = append(clientIds, r.Id)
-		recordIdToEmail[r.Id] = r.Email
-	}
-
-	emailsByInbound := map[int][]string{}
-	if len(clientIds) > 0 {
-		var mappings []model.ClientInbound
-		for _, batch := range chunkInts(clientIds, sqlInChunk) {
-			var rows []model.ClientInbound
-			if err := db.Where("client_id IN ?", batch).Find(&rows).Error; err != nil {
-				return result, false, err
-			}
-			mappings = append(mappings, rows...)
-		}
-		for _, m := range mappings {
-			email, ok := recordIdToEmail[m.ClientId]
-			if !ok {
-				continue
-			}
-			emailsByInbound[m.InboundId] = append(emailsByInbound[m.InboundId], email)
-		}
-	}
-
-	needRestart := false
-	for inboundId, ibEmails := range emailsByInbound {
-		ibResult := s.bulkDelInboundClients(inboundSvc, inboundId, ibEmails, recordsByEmail, false)
-		if ibResult.needRestart {
-			needRestart = true
-		}
-		for email, reason := range ibResult.perEmailSkipped {
-			if _, already := skippedReasons[email]; !already {
-				skippedReasons[email] = reason
-			}
-		}
-	}
-
-	successEmails := make([]string, 0, len(recordsByEmail))
-	successIds := make([]int, 0, len(recordsByEmail))
-	for email, rec := range recordsByEmail {
-		if _, skipped := skippedReasons[email]; skipped {
-			continue
-		}
-		successEmails = append(successEmails, email)
-		successIds = append(successIds, rec.Id)
-	}
-
-	if len(successIds) > 0 {
-		for _, batch := range chunkInts(successIds, sqlInChunk) {
-			if err := db.Where("client_id IN ?", batch).Delete(&model.ClientInbound{}).Error; err != nil {
-				return result, needRestart, err
-			}
-		}
-		if !keepTraffic && len(successEmails) > 0 {
-			for _, batch := range chunkStrings(successEmails, sqlInChunk) {
-				if err := db.Where("email IN ?", batch).Delete(&xray.ClientTraffic{}).Error; err != nil {
-					return result, needRestart, err
-				}
-				if err := db.Where("client_email IN ?", batch).Delete(&model.InboundClientIps{}).Error; err != nil {
-					return result, needRestart, err
-				}
-			}
-		}
-		for _, batch := range chunkInts(successIds, sqlInChunk) {
-			if err := db.Where("id IN ?", batch).Delete(&model.ClientRecord{}).Error; err != nil {
-				return result, needRestart, err
-			}
-		}
-	}
-
-	result.Deleted = len(successEmails)
-	for email, reason := range skippedReasons {
-		result.Skipped = append(result.Skipped, BulkDeleteReport{Email: email, Reason: reason})
-	}
-	return result, needRestart, nil
-}
-
-type bulkInboundDeleteResult struct {
-	perEmailSkipped map[string]string
-	needRestart     bool
-}
-
-// bulkDelInboundClients removes multiple clients from a single inbound's
-// settings JSON in one read-modify-write cycle, runs the xray runtime
-// RemoveUser/DeleteUser calls, and persists the inbound. The returned map
-// holds per-email failure reasons; emails not present in the map are
-// considered successful for this inbound.
-func (s *ClientService) bulkDelInboundClients(
-	inboundSvc *InboundService,
-	inboundId int,
-	emails []string,
-	records map[string]*model.ClientRecord,
-	keepTraffic bool,
-) bulkInboundDeleteResult {
-	res := bulkInboundDeleteResult{perEmailSkipped: map[string]string{}}
-
-	defer lockInbound(inboundId).Unlock()
-
-	oldInbound, err := inboundSvc.GetInbound(inboundId)
-	if err != nil {
-		logger.Error("Load Old Data Error")
-		for _, e := range emails {
-			res.perEmailSkipped[e] = err.Error()
-		}
-		return res
-	}
-
-	var settings map[string]any
-	if err := json.Unmarshal([]byte(oldInbound.Settings), &settings); err != nil {
-		for _, e := range emails {
-			res.perEmailSkipped[e] = err.Error()
-		}
-		return res
-	}
-
-	// Match by email — the client's stable identity (see Delete). Removes every
-	// entry carrying a wanted email, independent of credential drift.
-	wantedEmails := make(map[string]struct{}, len(emails))
-	for _, email := range emails {
-		if records[email] == nil {
-			res.perEmailSkipped[email] = "client not found"
-			continue
-		}
-		wantedEmails[email] = struct{}{}
-	}
-
-	interfaceClients, _ := settings["clients"].([]any)
-	newClients := make([]any, 0, len(interfaceClients))
-	foundEmails := map[string]bool{}
-	enableByEmail := map[string]bool{}
-	for _, client := range interfaceClients {
-		c, ok := client.(map[string]any)
-		if !ok {
-			newClients = append(newClients, client)
-			continue
-		}
-		em, _ := c["email"].(string)
-		if _, found := wantedEmails[em]; found && em != "" {
-			foundEmails[em] = true
-			en, _ := c["enable"].(bool)
-			enableByEmail[em] = en
-			continue
-		}
-		newClients = append(newClients, client)
-	}
-
-	for email := range wantedEmails {
-		if !foundEmails[email] {
-			res.perEmailSkipped[email] = "Client Not Found In Inbound"
-		}
-	}
-
-	db := database.GetDB()
-	newClients = compactOrphans(db, newClients)
-	if newClients == nil {
-		newClients = []any{}
-	}
-	settings["clients"] = newClients
-	newSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		for email := range foundEmails {
-			if _, skip := res.perEmailSkipped[email]; !skip {
-				res.perEmailSkipped[email] = err.Error()
-			}
-		}
-		return res
-	}
-	oldInbound.Settings = string(newSettings)
-
-	foundList := make([]string, 0, len(foundEmails))
-	for email := range foundEmails {
-		foundList = append(foundList, email)
-	}
-
-	notDepletedByEmail := map[string]bool{}
-	if len(foundList) > 0 {
-		type trafficRow struct {
-			Email  string
-			Enable bool
-		}
-		for _, batch := range chunkStrings(foundList, sqlInChunk) {
-			var rows []trafficRow
-			if err := db.Model(xray.ClientTraffic{}).
-				Where("email IN ?", batch).
-				Select("email, enable").
-				Scan(&rows).Error; err == nil {
-				for _, r := range rows {
-					notDepletedByEmail[r.Email] = r.Enable
-				}
-			}
-		}
-	}
-
-	var sharedSet map[string]bool
-	if !keepTraffic {
-		var sharedErr error
-		sharedSet, sharedErr = inboundSvc.emailsUsedByOtherInbounds(foundList, inboundId)
-		if sharedErr != nil {
-			for email := range foundEmails {
-				res.perEmailSkipped[email] = sharedErr.Error()
-				delete(foundEmails, email)
-			}
-			return res
-		}
-	}
-	if !keepTraffic {
-		purge := make([]string, 0, len(foundEmails))
-		for email := range foundEmails {
-			if !sharedSet[strings.ToLower(strings.TrimSpace(email))] {
-				purge = append(purge, email)
-			}
-		}
-		if len(purge) > 0 {
-			if delErr := inboundSvc.delClientIPsByEmails(db, purge); delErr != nil {
-				logger.Error("Error in delete client IPs")
-				for _, email := range purge {
-					res.perEmailSkipped[email] = delErr.Error()
-					delete(foundEmails, email)
-				}
-			} else if delErr := inboundSvc.delClientStatsByEmails(db, purge); delErr != nil {
-				logger.Error("Delete stats Data Error")
-				for _, email := range purge {
-					res.perEmailSkipped[email] = delErr.Error()
-					delete(foundEmails, email)
-				}
-			}
-		}
-	}
-
-	markDirty := false
-	if oldInbound.NodeID == nil {
-		rt, rterr := inboundSvc.runtimeFor(oldInbound)
-		if rterr != nil {
-			res.needRestart = true
-		} else {
-			for email := range foundEmails {
-				if !enableByEmail[email] || !notDepletedByEmail[email] {
-					continue
-				}
-				err1 := rt.RemoveUser(context.Background(), oldInbound, email)
-				if err1 == nil {
-					logger.Debug("Client deleted on", rt.Name(), ":", email)
-				} else if strings.Contains(err1.Error(), fmt.Sprintf("User %s not found.", email)) {
-					logger.Debug("User is already deleted. Nothing to do more...")
-				} else {
-					logger.Debug("Error in deleting client on", rt.Name(), ":", err1)
-					res.needRestart = true
-				}
-			}
-		}
-	} else {
-		rt, push, dirty, perr := inboundSvc.nodePushPlan(oldInbound)
-		if perr != nil {
-			for email := range foundEmails {
-				res.perEmailSkipped[email] = perr.Error()
-				delete(foundEmails, email)
-			}
-		} else {
-			if dirty {
-				markDirty = true
-			}
-			if push {
-				for email := range foundEmails {
-					if err1 := rt.DeleteUser(context.Background(), oldInbound, email); err1 != nil {
-						logger.Warning("Error in deleting client on", rt.Name(), ":", err1)
-						markDirty = true
-					}
-				}
-			}
-		}
-	}
-
-	txErr := db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Save(oldInbound).Error; err != nil {
-			return err
-		}
-		finalClients, err := inboundSvc.GetClients(oldInbound)
-		if err != nil {
-			return err
-		}
-		return s.SyncInbound(tx, inboundId, finalClients)
-	})
-	if txErr != nil {
-		for email := range foundEmails {
-			if _, skip := res.perEmailSkipped[email]; !skip {
-				res.perEmailSkipped[email] = txErr.Error()
-			}
-		}
-	} else if markDirty && oldInbound.NodeID != nil {
-		if dErr := (&NodeService{}).MarkNodeDirty(*oldInbound.NodeID); dErr != nil {
-			logger.Warning("mark node dirty failed:", dErr)
-		}
-	}
-
-	return res
-}
-
-// BulkCreateResult mirrors BulkAdjustResult for the create flow.
-type BulkCreateResult struct {
-	Created int                `json:"created"`
-	Skipped []BulkCreateReport `json:"skipped,omitempty"`
-}
-
-type BulkCreateReport struct {
-	Email  string `json:"email"`
-	Reason string `json:"reason"`
-}
-
-func (s *ClientService) BulkCreate(inboundSvc *InboundService, payloads []ClientCreatePayload) (BulkCreateResult, bool, error) {
-	result := BulkCreateResult{}
-	if len(payloads) == 0 {
-		return result, false, nil
-	}
-
-	skip := func(email, reason string) {
-		if strings.TrimSpace(email) == "" {
-			email = "(missing email)"
-		}
-		result.Skipped = append(result.Skipped, BulkCreateReport{Email: email, Reason: reason})
-	}
-
-	emailSubIDs, err := inboundSvc.getAllEmailSubIDs()
-	if err != nil {
-		emailSubIDs = nil
-	}
-
-	type prepared struct {
-		client     model.Client
-		inboundIds []int
-	}
-	prep := make([]prepared, 0, len(payloads))
-	emails := make([]string, 0, len(payloads))
-	subIDs := make([]string, 0, len(payloads))
-	seenEmail := make(map[string]struct{}, len(payloads))
-	seenSubID := make(map[string]string, len(payloads))
-
-	for i := range payloads {
-		client := payloads[i].Client
-		email := strings.TrimSpace(client.Email)
-		if email == "" {
-			skip("", "client email is required")
-			continue
-		}
-		if verr := validateClientEmail(email); verr != nil {
-			skip(email, verr.Error())
-			continue
-		}
-		if verr := validateClientSubID(client.SubID); verr != nil {
-			skip(email, verr.Error())
-			continue
-		}
-		if len(payloads[i].InboundIds) == 0 {
-			skip(email, "at least one inbound is required")
-			continue
-		}
-
-		client.Email = email
-		if client.SubID == "" {
-			client.SubID = uuid.NewString()
-		}
-		if !client.Enable {
-			client.Enable = true
-		}
-		now := time.Now().UnixMilli()
-		if client.CreatedAt == 0 {
-			client.CreatedAt = now
-		}
-		client.UpdatedAt = now
-
-		le := strings.ToLower(email)
-		if _, dup := seenEmail[le]; dup {
-			skip(email, "email already in use: "+email)
-			continue
-		}
-		if owner, ok := seenSubID[client.SubID]; ok && owner != le {
-			skip(email, "subId already in use: "+client.SubID)
-			continue
-		}
-		seenEmail[le] = struct{}{}
-		seenSubID[client.SubID] = le
-
-		prep = append(prep, prepared{client: client, inboundIds: payloads[i].InboundIds})
-		emails = append(emails, email)
-		subIDs = append(subIDs, client.SubID)
-	}
-
-	if len(prep) == 0 {
-		return result, false, nil
-	}
-
-	db := database.GetDB()
-	const lookupChunk = 400
-	existingEmailSub := make(map[string]string, len(emails))
-	for start := 0; start < len(emails); start += lookupChunk {
-		end := min(start+lookupChunk, len(emails))
-		var rows []model.ClientRecord
-		if e := db.Where("email IN ?", emails[start:end]).Find(&rows).Error; e != nil {
-			return result, false, e
-		}
-		for i := range rows {
-			existingEmailSub[strings.ToLower(rows[i].Email)] = rows[i].SubID
-		}
-	}
-	existingSubOwner := make(map[string]string, len(subIDs))
-	for start := 0; start < len(subIDs); start += lookupChunk {
-		end := min(start+lookupChunk, len(subIDs))
-		var rows []model.ClientRecord
-		if e := db.Where("sub_id IN ?", subIDs[start:end]).Find(&rows).Error; e != nil {
-			return result, false, e
-		}
-		for i := range rows {
-			existingSubOwner[rows[i].SubID] = strings.ToLower(rows[i].Email)
-		}
-	}
-
-	inboundCache := make(map[int]*model.Inbound)
-	getIb := func(id int) (*model.Inbound, error) {
-		if ib, ok := inboundCache[id]; ok {
-			return ib, nil
-		}
-		ib, e := inboundSvc.GetInbound(id)
-		if e != nil {
-			return nil, e
-		}
-		inboundCache[id] = ib
-		return ib, nil
-	}
-
-	byInbound := make(map[int][]model.Client)
-	idxByInbound := make(map[int][]int)
-	inboundOrder := make([]int, 0)
-	failed := make([]bool, len(prep))
-	reason := make([]string, len(prep))
-
-	for idx := range prep {
-		le := strings.ToLower(prep[idx].client.Email)
-		if existSub, ok := existingEmailSub[le]; ok && existSub != prep[idx].client.SubID {
-			failed[idx] = true
-			reason[idx] = "email already in use: " + prep[idx].client.Email
-			continue
-		}
-		if owner, ok := existingSubOwner[prep[idx].client.SubID]; ok && owner != le {
-			failed[idx] = true
-			reason[idx] = "subId already in use: " + prep[idx].client.SubID
-			continue
-		}
-
-		ok := true
-		for _, ibId := range prep[idx].inboundIds {
-			ib, e := getIb(ibId)
-			if e != nil {
-				failed[idx] = true
-				reason[idx] = e.Error()
-				ok = false
-				break
-			}
-			if e := s.fillProtocolDefaults(&prep[idx].client, ib); e != nil {
-				failed[idx] = true
-				reason[idx] = e.Error()
-				ok = false
-				break
-			}
-		}
-		if !ok {
-			continue
-		}
-		for _, ibId := range prep[idx].inboundIds {
-			ib, _ := getIb(ibId)
-			if _, seen := byInbound[ibId]; !seen {
-				inboundOrder = append(inboundOrder, ibId)
-			}
-			byInbound[ibId] = append(byInbound[ibId], clientWithInboundFlow(prep[idx].client, ib))
-			idxByInbound[ibId] = append(idxByInbound[ibId], idx)
-		}
-	}
-
-	needRestart := false
-	for _, ibId := range inboundOrder {
-		payload, e := json.Marshal(map[string][]model.Client{"clients": byInbound[ibId]})
-		if e == nil {
-			var nr bool
-			nr, e = s.addInboundClient(inboundSvc, &model.Inbound{Id: ibId, Settings: string(payload)}, emailSubIDs)
-			if e == nil && nr {
-				needRestart = true
-			}
-		}
-		if e != nil {
-			for _, idx := range idxByInbound[ibId] {
-				failed[idx] = true
-				if reason[idx] == "" {
-					reason[idx] = e.Error()
-				}
-			}
-		}
-	}
-
-	for idx := range prep {
-		if failed[idx] {
-			skip(prep[idx].client.Email, reason[idx])
-		} else {
-			result.Created++
-		}
-	}
-	return result, needRestart, nil
-}
-
-func (s *ClientService) DelDepleted(inboundSvc *InboundService) (int, bool, error) {
-	db := database.GetDB()
-	now := time.Now().UnixMilli()
-	depletedClause := "reset = 0 and ((total > 0 and up + down >= total) or (expiry_time > 0 and expiry_time <= ?))"
-
-	var rows []xray.ClientTraffic
-	if err := db.Where(depletedClause, now).Find(&rows).Error; err != nil {
-		return 0, false, err
-	}
-	if len(rows) == 0 {
-		return 0, false, nil
-	}
-
-	seen := make(map[string]struct{}, len(rows))
-	emails := make([]string, 0, len(rows))
-	for _, r := range rows {
-		if r.Email == "" {
-			continue
-		}
-		if _, ok := seen[r.Email]; ok {
-			continue
-		}
-		seen[r.Email] = struct{}{}
-		emails = append(emails, r.Email)
-	}
-	if len(emails) == 0 {
-		return 0, false, nil
-	}
-
-	res, needRestart, err := s.BulkDelete(inboundSvc, emails, false)
-	if err != nil {
-		return res.Deleted, needRestart, err
-	}
-	return res.Deleted, needRestart, nil
-}
-
-func (s *ClientService) ResetAllClientTraffics(inboundSvc *InboundService, id int) error {
-	return submitTrafficWrite(func() error {
-		return s.resetAllClientTrafficsLocked(id)
-	})
-}
-
-func (s *ClientService) resetAllClientTrafficsLocked(id int) error {
-	db := database.GetDB()
-	now := time.Now().Unix() * 1000
-
-	if err := db.Transaction(func(tx *gorm.DB) error {
-		whereText := "inbound_id "
-		if id == -1 {
-			whereText += " > ?"
-		} else {
-			whereText += " = ?"
-		}
-
-		result := tx.Model(xray.ClientTraffic{}).
-			Where(whereText, id).
-			Updates(map[string]any{"enable": true, "up": 0, "down": 0})
-
-		if result.Error != nil {
-			return result.Error
-		}
-
-		inboundWhereText := "id "
-		if id == -1 {
-			inboundWhereText += " > ?"
-		} else {
-			inboundWhereText += " = ?"
-		}
-
-		result = tx.Model(model.Inbound{}).
-			Where(inboundWhereText, id).
-			Update("last_traffic_reset_time", now)
-
-		return result.Error
-	}); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (s *ClientService) ResetAllTraffics() (bool, error) {
-	res := database.GetDB().Model(&xray.ClientTraffic{}).
-		Where("1 = 1").
-		Updates(map[string]any{"up": 0, "down": 0})
-	if res.Error != nil {
-		return false, res.Error
-	}
-	return res.RowsAffected > 0, nil
-}
-
-func (s *ClientService) Detach(inboundSvc *InboundService, id int, inboundIds []int) (bool, error) {
-	existing, err := s.GetByID(id)
-	if err != nil {
-		return false, err
-	}
-	currentIds, err := s.GetInboundIdsForRecord(id)
-	if err != nil {
-		return false, err
-	}
-	have := make(map[int]struct{}, len(currentIds))
-	for _, x := range currentIds {
-		have[x] = struct{}{}
-	}
-
-	needRestart := false
-	for _, ibId := range inboundIds {
-		if _, attached := have[ibId]; !attached {
-			continue
-		}
-		if _, getErr := inboundSvc.GetInbound(ibId); getErr != nil {
-			return needRestart, getErr
-		}
-		// Detach by email — the client's stable identity (see Delete).
-		if existing.Email == "" {
-			continue
-		}
-		nr, delErr := s.DelInboundClientByEmail(inboundSvc, ibId, existing.Email, true)
-		if delErr != nil {
-			if errors.Is(delErr, ErrClientNotInInbound) {
-				continue
-			}
-			return needRestart, delErr
-		}
-		if nr {
-			needRestart = true
-		}
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) checkEmailsExistForClients(inboundSvc *InboundService, clients []model.Client, emailSubIDs map[string]string) (string, error) {
-	if emailSubIDs == nil {
-		var err error
-		emailSubIDs, err = inboundSvc.getAllEmailSubIDs()
-		if err != nil {
-			return "", err
-		}
-	}
-	seen := make(map[string]string, len(clients))
-	for _, client := range clients {
-		if client.Email == "" {
-			continue
-		}
-		key := strings.ToLower(client.Email)
-		if prev, ok := seen[key]; ok {
-			if prev != client.SubID || client.SubID == "" {
-				return client.Email, nil
-			}
-			continue
-		}
-		seen[key] = client.SubID
-		if existingSub, ok := emailSubIDs[key]; ok {
-			if client.SubID == "" || existingSub == "" || existingSub != client.SubID {
-				return client.Email, nil
-			}
-		}
-	}
-	return "", nil
-}
-
-func (s *ClientService) AddInboundClient(inboundSvc *InboundService, data *model.Inbound) (bool, error) {
-	return s.addInboundClient(inboundSvc, data, nil)
-}
-
-// addInboundClient is AddInboundClient with an optional precomputed email→subId
-// map. Bulk callers pass a single snapshot so the global getAllEmailSubIDs scan
-// runs once for the whole batch instead of once per target inbound; a nil map
-// makes it compute its own (the single-add path).
-func (s *ClientService) addInboundClient(inboundSvc *InboundService, data *model.Inbound, emailSubIDs map[string]string) (bool, error) {
-	defer lockInbound(data.Id).Unlock()
-
-	clients, err := inboundSvc.GetClients(data)
-	if err != nil {
-		return false, err
-	}
-
-	var settings map[string]any
-	err = json.Unmarshal([]byte(data.Settings), &settings)
-	if err != nil {
-		return false, err
-	}
-
-	interfaceClients := settings["clients"].([]any)
-	nowTs := time.Now().Unix() * 1000
-	for i := range interfaceClients {
-		if cm, ok := interfaceClients[i].(map[string]any); ok {
-			if _, ok2 := cm["created_at"]; !ok2 {
-				cm["created_at"] = nowTs
-			}
-			cm["updated_at"] = nowTs
-			existingSub, _ := cm["subId"].(string)
-			if strings.TrimSpace(existingSub) == "" {
-				cm["subId"] = random.NumLower(16)
-			}
-			interfaceClients[i] = cm
-		}
-	}
-	existEmail, err := s.checkEmailsExistForClients(inboundSvc, clients, emailSubIDs)
-	if err != nil {
-		return false, err
-	}
-	if existEmail != "" {
-		return false, common.NewError("Duplicate email:", existEmail)
-	}
-
-	oldInbound, err := inboundSvc.GetInbound(data.Id)
-	if err != nil {
-		return false, err
-	}
-
-	for _, client := range clients {
-		if strings.TrimSpace(client.Email) == "" {
-			return false, common.NewError("client email is required")
-		}
-		switch oldInbound.Protocol {
-		case "trojan":
-			if client.Password == "" {
-				return false, common.NewError("empty client ID")
-			}
-		case "shadowsocks":
-			if client.Email == "" {
-				return false, common.NewError("empty client ID")
-			}
-		case "hysteria":
-			if client.Auth == "" {
-				return false, common.NewError("empty client ID")
-			}
-		default:
-			if client.ID == "" {
-				return false, common.NewError("empty client ID")
-			}
-		}
-	}
-
-	var oldSettings map[string]any
-	err = json.Unmarshal([]byte(oldInbound.Settings), &oldSettings)
-	if err != nil {
-		return false, err
-	}
-
-	if oldInbound.Protocol == model.Shadowsocks {
-		applyShadowsocksClientMethod(interfaceClients, oldSettings)
-	}
-
-	oldClients := oldSettings["clients"].([]any)
-	oldClients = compactOrphans(database.GetDB(), oldClients)
-	oldClients = append(oldClients, interfaceClients...)
-
-	oldSettings["clients"] = oldClients
-
-	newSettings, err := json.MarshalIndent(oldSettings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-
-	oldInbound.Settings = string(newSettings)
-
-	db := database.GetDB()
-	tx := db.Begin()
-
-	markDirty := false
-	defer func() {
-		if err != nil {
-			tx.Rollback()
-			return
-		}
-		tx.Commit()
-		if markDirty && oldInbound.NodeID != nil {
-			if dErr := (&NodeService{}).MarkNodeDirty(*oldInbound.NodeID); dErr != nil {
-				logger.Warning("mark node dirty failed:", dErr)
-			}
-		}
-	}()
-
-	needRestart := false
-	rt, push, dirty, perr := inboundSvc.nodePushPlan(oldInbound)
-	if perr != nil {
-		err = perr
-		return false, err
-	}
-	if dirty {
-		markDirty = true
-	}
-	if oldInbound.NodeID == nil {
-		if !push {
-			needRestart = true
-		} else {
-			for _, client := range clients {
-				if len(client.Email) == 0 {
-					needRestart = true
-					continue
-				}
-				inboundSvc.AddClientStat(tx, data.Id, &client)
-				if !client.Enable {
-					continue
-				}
-				cipher := ""
-				if oldInbound.Protocol == "shadowsocks" {
-					cipher = oldSettings["method"].(string)
-				}
-				err1 := rt.AddUser(context.Background(), oldInbound, map[string]any{
-					"email":    client.Email,
-					"id":       client.ID,
-					"auth":     client.Auth,
-					"security": client.Security,
-					"flow":     client.Flow,
-					"password": client.Password,
-					"cipher":   cipher,
-				})
-				if err1 == nil {
-					logger.Debug("Client added on", rt.Name(), ":", client.Email)
-				} else {
-					logger.Debug("Error in adding client on", rt.Name(), ":", err1)
-					needRestart = true
-				}
-			}
-		}
-	} else {
-		for _, client := range clients {
-			if len(client.Email) > 0 {
-				inboundSvc.AddClientStat(tx, data.Id, &client)
-			}
-			if push {
-				if err1 := rt.AddClient(context.Background(), oldInbound, client); err1 != nil {
-					logger.Warning("Error in adding client on", rt.Name(), ":", err1)
-					markDirty = true
-					push = false
-				}
-			}
-		}
-	}
-
-	if err = tx.Save(oldInbound).Error; err != nil {
-		return false, err
-	}
-	finalClients, gcErr := inboundSvc.GetClients(oldInbound)
-	if gcErr != nil {
-		err = gcErr
-		return false, err
-	}
-	if err = s.SyncInbound(tx, oldInbound.Id, finalClients); err != nil {
-		return false, err
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) UpdateInboundClient(inboundSvc *InboundService, data *model.Inbound, oldEmail string) (bool, error) {
-	defer lockInbound(data.Id).Unlock()
-
-	clients, err := inboundSvc.GetClients(data)
-	if err != nil {
-		return false, err
-	}
-
-	var settings map[string]any
-	err = json.Unmarshal([]byte(data.Settings), &settings)
-	if err != nil {
-		return false, err
-	}
-
-	interfaceClients := settings["clients"].([]any)
-
-	oldInbound, err := inboundSvc.GetInbound(data.Id)
-	if err != nil {
-		return false, err
-	}
-
-	oldClients, err := inboundSvc.GetClients(oldInbound)
-	if err != nil {
-		return false, err
-	}
-
-	newClientId := ""
-	switch oldInbound.Protocol {
-	case "trojan":
-		newClientId = clients[0].Password
-	case "shadowsocks":
-		newClientId = clients[0].Email
-	case "hysteria":
-		newClientId = clients[0].Auth
-	default:
-		newClientId = clients[0].ID
-	}
-
-	// Locate the client to replace by email — the client's stable identity.
-	// Credentials (uuid/password/auth) can drift from the inbound JSON, so they
-	// are never used for matching.
-	clientIndex := -1
-	for index, oldClient := range oldClients {
-		if strings.EqualFold(oldClient.Email, oldEmail) {
-			oldEmail = oldClient.Email
-			clientIndex = index
-			break
-		}
-	}
-
-	if newClientId == "" || clientIndex == -1 {
-		return false, common.NewError("empty client ID")
-	}
-	if strings.TrimSpace(clients[0].Email) == "" {
-		return false, common.NewError("client email is required")
-	}
-
-	if clients[0].Email != oldEmail {
-		existEmail, err := s.checkEmailsExistForClients(inboundSvc, clients, nil)
-		if err != nil {
-			return false, err
-		}
-		if existEmail != "" {
-			return false, common.NewError("Duplicate email:", existEmail)
-		}
-	}
-
-	var oldSettings map[string]any
-	err = json.Unmarshal([]byte(oldInbound.Settings), &oldSettings)
-	if err != nil {
-		return false, err
-	}
-	settingsClients := oldSettings["clients"].([]any)
-	var preservedCreated any
-	var preservedSubID string
-	if clientIndex >= 0 && clientIndex < len(settingsClients) {
-		if oldMap, ok := settingsClients[clientIndex].(map[string]any); ok {
-			if v, ok2 := oldMap["created_at"]; ok2 {
-				preservedCreated = v
-			}
-			preservedSubID, _ = oldMap["subId"].(string)
-		}
-	}
-	if len(interfaceClients) > 0 {
-		if newMap, ok := interfaceClients[0].(map[string]any); ok {
-			if preservedCreated == nil {
-				preservedCreated = time.Now().Unix() * 1000
-			}
-			newMap["created_at"] = preservedCreated
-			newMap["updated_at"] = time.Now().Unix() * 1000
-			newSub, _ := newMap["subId"].(string)
-			if strings.TrimSpace(newSub) == "" {
-				if strings.TrimSpace(preservedSubID) != "" {
-					newMap["subId"] = preservedSubID
-				} else {
-					newMap["subId"] = random.NumLower(16)
-				}
-			}
-			interfaceClients[0] = newMap
-		}
-	}
-	if oldInbound.Protocol == model.Shadowsocks {
-		applyShadowsocksClientMethod(interfaceClients, oldSettings)
-	}
-	settingsClients[clientIndex] = interfaceClients[0]
-	oldSettings["clients"] = settingsClients
-
-	if oldInbound.Protocol == model.VLESS {
-		hasVisionFlow := false
-		for _, c := range settingsClients {
-			cm, ok := c.(map[string]any)
-			if !ok {
-				continue
-			}
-			if flow, _ := cm["flow"].(string); flow == "xtls-rprx-vision" {
-				hasVisionFlow = true
-				break
-			}
-		}
-		if !hasVisionFlow {
-			delete(oldSettings, "testseed")
-		}
-	}
-
-	newSettings, err := json.MarshalIndent(oldSettings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-
-	oldInbound.Settings = string(newSettings)
-	db := database.GetDB()
-	tx := db.Begin()
-
-	markDirty := false
-	defer func() {
-		if err != nil {
-			tx.Rollback()
-			return
-		}
-		tx.Commit()
-		if markDirty && oldInbound.NodeID != nil {
-			if dErr := (&NodeService{}).MarkNodeDirty(*oldInbound.NodeID); dErr != nil {
-				logger.Warning("mark node dirty failed:", dErr)
-			}
-		}
-	}()
-
-	if len(clients[0].Email) > 0 {
-		if len(oldEmail) > 0 {
-			emailUnchanged := strings.EqualFold(oldEmail, clients[0].Email)
-			targetExists := int64(0)
-			if !emailUnchanged {
-				if err = tx.Model(xray.ClientTraffic{}).Where("email = ?", clients[0].Email).Count(&targetExists).Error; err != nil {
-					return false, err
-				}
-			}
-			if emailUnchanged || targetExists == 0 {
-				err = inboundSvc.UpdateClientStat(tx, oldEmail, &clients[0])
-				if err != nil {
-					return false, err
-				}
-				err = inboundSvc.UpdateClientIPs(tx, oldEmail, clients[0].Email)
-				if err != nil {
-					return false, err
-				}
-			} else {
-				stillUsed, sErr := inboundSvc.emailUsedByOtherInbounds(oldEmail, data.Id)
-				if sErr != nil {
-					return false, sErr
-				}
-				if !stillUsed {
-					if err = inboundSvc.DelClientStat(tx, oldEmail); err != nil {
-						return false, err
-					}
-					if err = inboundSvc.DelClientIPs(tx, oldEmail); err != nil {
-						return false, err
-					}
-				}
-				if err = inboundSvc.UpdateClientStat(tx, clients[0].Email, &clients[0]); err != nil {
-					return false, err
-				}
-			}
-		} else {
-			inboundSvc.AddClientStat(tx, data.Id, &clients[0])
-		}
-	} else {
-		stillUsed, err := inboundSvc.emailUsedByOtherInbounds(oldEmail, data.Id)
-		if err != nil {
-			return false, err
-		}
-		if !stillUsed {
-			err = inboundSvc.DelClientStat(tx, oldEmail)
-			if err != nil {
-				return false, err
-			}
-			err = inboundSvc.DelClientIPs(tx, oldEmail)
-			if err != nil {
-				return false, err
-			}
-		}
-	}
-	needRestart := false
-	if len(oldEmail) > 0 {
-		rt, push, dirty, perr := inboundSvc.nodePushPlan(oldInbound)
-		if perr != nil {
-			err = perr
-			return false, err
-		}
-		if dirty {
-			markDirty = true
-		}
-		if oldInbound.NodeID == nil {
-			if !push {
-				needRestart = true
-			} else {
-				if oldClients[clientIndex].Enable {
-					err1 := rt.RemoveUser(context.Background(), oldInbound, oldEmail)
-					if err1 == nil {
-						logger.Debug("Old client deleted on", rt.Name(), ":", oldEmail)
-					} else if strings.Contains(err1.Error(), fmt.Sprintf("User %s not found.", oldEmail)) {
-						logger.Debug("User is already deleted. Nothing to do more...")
-					} else {
-						logger.Debug("Error in deleting client on", rt.Name(), ":", err1)
-						needRestart = true
-					}
-				}
-				if clients[0].Enable {
-					cipher := ""
-					if oldInbound.Protocol == "shadowsocks" {
-						cipher = oldSettings["method"].(string)
-					}
-					err1 := rt.AddUser(context.Background(), oldInbound, map[string]any{
-						"email":    clients[0].Email,
-						"id":       clients[0].ID,
-						"security": clients[0].Security,
-						"flow":     clients[0].Flow,
-						"auth":     clients[0].Auth,
-						"password": clients[0].Password,
-						"cipher":   cipher,
-					})
-					if err1 == nil {
-						logger.Debug("Client edited on", rt.Name(), ":", clients[0].Email)
-					} else {
-						logger.Debug("Error in adding client on", rt.Name(), ":", err1)
-						needRestart = true
-					}
-				}
-			}
-		} else if push {
-			if err1 := rt.UpdateUser(context.Background(), oldInbound, oldEmail, clients[0]); err1 != nil {
-				logger.Warning("Error in updating client on", rt.Name(), ":", err1)
-				markDirty = true
-			}
-		}
-	} else {
-		logger.Debug("Client old email not found")
-		needRestart = true
-	}
-	if err = tx.Save(oldInbound).Error; err != nil {
-		return false, err
-	}
-	finalClients, gcErr := inboundSvc.GetClients(oldInbound)
-	if gcErr != nil {
-		err = gcErr
-		return false, err
-	}
-	if err = s.SyncInbound(tx, oldInbound.Id, finalClients); err != nil {
-		return false, err
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) DelInboundClientByEmail(inboundSvc *InboundService, inboundId int, email string, keepTraffic bool) (bool, error) {
-	defer lockInbound(inboundId).Unlock()
-
-	oldInbound, err := inboundSvc.GetInbound(inboundId)
-	if err != nil {
-		logger.Error("Load Old Data Error")
-		return false, err
-	}
-
-	var settings map[string]any
-	if err := json.Unmarshal([]byte(oldInbound.Settings), &settings); err != nil {
-		return false, err
-	}
-
-	interfaceClients, ok := settings["clients"].([]any)
-	if !ok {
-		return false, common.NewError("invalid clients format in inbound settings")
-	}
-
-	var newClients []any
-	needApiDel := false
-	found := false
-
-	for _, client := range interfaceClients {
-		c, ok := client.(map[string]any)
-		if !ok {
-			continue
-		}
-		if cEmail, ok := c["email"].(string); ok && cEmail == email {
-			found = true
-			needApiDel, _ = c["enable"].(bool)
-		} else {
-			newClients = append(newClients, client)
-		}
-	}
-
-	if !found {
-		return false, fmt.Errorf("%w for email: %s", ErrClientNotInInbound, email)
-	}
-	db := database.GetDB()
-	newClients = compactOrphans(db, newClients)
-	if newClients == nil {
-		newClients = []any{}
-	}
-	settings["clients"] = newClients
-	newSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-
-	oldInbound.Settings = string(newSettings)
-
-	emailShared, err := inboundSvc.emailUsedByOtherInbounds(email, inboundId)
-	if err != nil {
-		return false, err
-	}
-
-	if !emailShared && !keepTraffic {
-		if err := inboundSvc.DelClientIPs(db, email); err != nil {
-			logger.Error("Error in delete client IPs")
-			return false, err
-		}
-	}
-
-	needRestart := false
-	markDirty := false
-
-	if len(email) > 0 && !emailShared {
-		if !keepTraffic {
-			traffic, err := inboundSvc.GetClientTrafficByEmail(email)
-			if err != nil {
-				return false, err
-			}
-			if traffic != nil {
-				if err := inboundSvc.DelClientStat(db, email); err != nil {
-					logger.Error("Delete stats Data Error")
-					return false, err
-				}
-			}
-		}
-
-		if needApiDel {
-			rt, push, dirty, perr := inboundSvc.nodePushPlan(oldInbound)
-			if perr != nil {
-				return false, perr
-			}
-			if dirty {
-				markDirty = true
-			}
-			if oldInbound.NodeID == nil {
-				if !push {
-					needRestart = true
-				} else if err1 := rt.RemoveUser(context.Background(), oldInbound, email); err1 == nil {
-					logger.Debug("Client deleted on", rt.Name(), ":", email)
-					needRestart = false
-				} else if strings.Contains(err1.Error(), fmt.Sprintf("User %s not found.", email)) {
-					logger.Debug("User is already deleted. Nothing to do more...")
-				} else {
-					logger.Debug("Error in deleting client on", rt.Name(), ":", email)
-					needRestart = true
-				}
-			} else if push {
-				if err1 := rt.DeleteUser(context.Background(), oldInbound, email); err1 != nil {
-					logger.Warning("Error in deleting client on", rt.Name(), ":", err1)
-					markDirty = true
-				}
-			}
-		}
-	}
-
-	if err := db.Save(oldInbound).Error; err != nil {
-		return false, err
-	}
-	finalClients, gcErr := inboundSvc.GetClients(oldInbound)
-	if gcErr != nil {
-		return false, gcErr
-	}
-	if err := s.SyncInbound(db, inboundId, finalClients); err != nil {
-		return false, err
-	}
-	if markDirty && oldInbound.NodeID != nil {
-		if dErr := (&NodeService{}).MarkNodeDirty(*oldInbound.NodeID); dErr != nil {
-			logger.Warning("mark node dirty failed:", dErr)
-		}
-	}
-	return needRestart, nil
-}
-
-func (s *ClientService) SetClientTelegramUserID(inboundSvc *InboundService, trafficId int, tgId int64) (bool, error) {
-	traffic, inbound, err := inboundSvc.GetClientInboundByTrafficID(trafficId)
-	if err != nil {
-		return false, err
-	}
-	if inbound == nil {
-		return false, common.NewError("Inbound Not Found For Traffic ID:", trafficId)
-	}
-
-	clientEmail := traffic.Email
-
-	oldClients, err := inboundSvc.GetClients(inbound)
-	if err != nil {
-		return false, err
-	}
-
-	found := false
-	for _, oldClient := range oldClients {
-		if oldClient.Email == clientEmail {
-			found = true
-			break
-		}
-	}
-
-	if !found {
-		return false, common.NewError("Client Not Found For Email:", clientEmail)
-	}
-
-	var settings map[string]any
-	err = json.Unmarshal([]byte(inbound.Settings), &settings)
-	if err != nil {
-		return false, err
-	}
-	clients := settings["clients"].([]any)
-	var newClients []any
-	for client_index := range clients {
-		c := clients[client_index].(map[string]any)
-		if c["email"] == clientEmail {
-			c["tgId"] = tgId
-			c["updated_at"] = time.Now().Unix() * 1000
-			newClients = append(newClients, any(c))
-		}
-	}
-	settings["clients"] = newClients
-	modifiedSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-	inbound.Settings = string(modifiedSettings)
-	needRestart, err := s.UpdateInboundClient(inboundSvc, inbound, clientEmail)
-	return needRestart, err
-}
-
-func (s *ClientService) checkIsEnabledByEmail(inboundSvc *InboundService, clientEmail string) (bool, error) {
-	_, inbound, err := inboundSvc.GetClientInboundByEmail(clientEmail)
-	if err != nil {
-		return false, err
-	}
-	if inbound == nil {
-		return false, common.NewError("Inbound Not Found For Email:", clientEmail)
-	}
-
-	clients, err := inboundSvc.GetClients(inbound)
-	if err != nil {
-		return false, err
-	}
-
-	isEnable := false
-
-	for _, client := range clients {
-		if client.Email == clientEmail {
-			isEnable = client.Enable
-			break
-		}
-	}
-
-	return isEnable, err
-}
-
-func (s *ClientService) ToggleClientEnableByEmail(inboundSvc *InboundService, clientEmail string) (bool, bool, error) {
-	_, inbound, err := inboundSvc.GetClientInboundByEmail(clientEmail)
-	if err != nil {
-		return false, false, err
-	}
-	if inbound == nil {
-		return false, false, common.NewError("Inbound Not Found For Email:", clientEmail)
-	}
-
-	oldClients, err := inboundSvc.GetClients(inbound)
-	if err != nil {
-		return false, false, err
-	}
-
-	found := false
-	clientOldEnabled := false
-
-	for _, oldClient := range oldClients {
-		if oldClient.Email == clientEmail {
-			found = true
-			clientOldEnabled = oldClient.Enable
-			break
-		}
-	}
-
-	if !found {
-		return false, false, common.NewError("Client Not Found For Email:", clientEmail)
-	}
-
-	var settings map[string]any
-	err = json.Unmarshal([]byte(inbound.Settings), &settings)
-	if err != nil {
-		return false, false, err
-	}
-	clients := settings["clients"].([]any)
-	var newClients []any
-	for client_index := range clients {
-		c := clients[client_index].(map[string]any)
-		if c["email"] == clientEmail {
-			c["enable"] = !clientOldEnabled
-			c["updated_at"] = time.Now().Unix() * 1000
-			newClients = append(newClients, any(c))
-		}
-	}
-	settings["clients"] = newClients
-	modifiedSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return false, false, err
-	}
-	inbound.Settings = string(modifiedSettings)
-
-	needRestart, err := s.UpdateInboundClient(inboundSvc, inbound, clientEmail)
-	if err != nil {
-		return false, needRestart, err
-	}
-
-	return !clientOldEnabled, needRestart, nil
-}
-
-func (s *ClientService) SetClientEnableByEmail(inboundSvc *InboundService, clientEmail string, enable bool) (bool, bool, error) {
-	current, err := s.checkIsEnabledByEmail(inboundSvc, clientEmail)
-	if err != nil {
-		return false, false, err
-	}
-	if current == enable {
-		return false, false, nil
-	}
-	newEnabled, needRestart, err := s.ToggleClientEnableByEmail(inboundSvc, clientEmail)
-	if err != nil {
-		return false, needRestart, err
-	}
-	return newEnabled == enable, needRestart, nil
-}
-
-func (s *ClientService) ResetClientIpLimitByEmail(inboundSvc *InboundService, clientEmail string, count int) (bool, error) {
-	_, inbound, err := inboundSvc.GetClientInboundByEmail(clientEmail)
-	if err != nil {
-		return false, err
-	}
-	if inbound == nil {
-		return false, common.NewError("Inbound Not Found For Email:", clientEmail)
-	}
-
-	oldClients, err := inboundSvc.GetClients(inbound)
-	if err != nil {
-		return false, err
-	}
-
-	found := false
-	for _, oldClient := range oldClients {
-		if oldClient.Email == clientEmail {
-			found = true
-			break
-		}
-	}
-
-	if !found {
-		return false, common.NewError("Client Not Found For Email:", clientEmail)
-	}
-
-	var settings map[string]any
-	err = json.Unmarshal([]byte(inbound.Settings), &settings)
-	if err != nil {
-		return false, err
-	}
-	clients := settings["clients"].([]any)
-	var newClients []any
-	for client_index := range clients {
-		c := clients[client_index].(map[string]any)
-		if c["email"] == clientEmail {
-			c["limitIp"] = count
-			c["updated_at"] = time.Now().Unix() * 1000
-			newClients = append(newClients, any(c))
-		}
-	}
-	settings["clients"] = newClients
-	modifiedSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-	inbound.Settings = string(modifiedSettings)
-	needRestart, err := s.UpdateInboundClient(inboundSvc, inbound, clientEmail)
-	return needRestart, err
-}
-
-func (s *ClientService) ResetClientExpiryTimeByEmail(inboundSvc *InboundService, clientEmail string, expiry_time int64) (bool, error) {
-	_, inbound, err := inboundSvc.GetClientInboundByEmail(clientEmail)
-	if err != nil {
-		return false, err
-	}
-	if inbound == nil {
-		return false, common.NewError("Inbound Not Found For Email:", clientEmail)
-	}
-
-	oldClients, err := inboundSvc.GetClients(inbound)
-	if err != nil {
-		return false, err
-	}
-
-	found := false
-	for _, oldClient := range oldClients {
-		if oldClient.Email == clientEmail {
-			found = true
-			break
-		}
-	}
-
-	if !found {
-		return false, common.NewError("Client Not Found For Email:", clientEmail)
-	}
-
-	var settings map[string]any
-	err = json.Unmarshal([]byte(inbound.Settings), &settings)
-	if err != nil {
-		return false, err
-	}
-	clients := settings["clients"].([]any)
-	var newClients []any
-	for client_index := range clients {
-		c := clients[client_index].(map[string]any)
-		if c["email"] == clientEmail {
-			c["expiryTime"] = expiry_time
-			c["updated_at"] = time.Now().Unix() * 1000
-			newClients = append(newClients, any(c))
-		}
-	}
-	settings["clients"] = newClients
-	modifiedSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-	inbound.Settings = string(modifiedSettings)
-	needRestart, err := s.UpdateInboundClient(inboundSvc, inbound, clientEmail)
-	return needRestart, err
-}
-
-func (s *ClientService) ResetClientTrafficLimitByEmail(inboundSvc *InboundService, clientEmail string, totalGB int) (bool, error) {
-	if totalGB < 0 {
-		return false, common.NewError("totalGB must be >= 0")
-	}
-	_, inbound, err := inboundSvc.GetClientInboundByEmail(clientEmail)
-	if err != nil {
-		return false, err
-	}
-	if inbound == nil {
-		return false, common.NewError("Inbound Not Found For Email:", clientEmail)
-	}
-
-	oldClients, err := inboundSvc.GetClients(inbound)
-	if err != nil {
-		return false, err
-	}
-
-	found := false
-	for _, oldClient := range oldClients {
-		if oldClient.Email == clientEmail {
-			found = true
-			break
-		}
-	}
-
-	if !found {
-		return false, common.NewError("Client Not Found For Email:", clientEmail)
-	}
-
-	var settings map[string]any
-	err = json.Unmarshal([]byte(inbound.Settings), &settings)
-	if err != nil {
-		return false, err
-	}
-	clients := settings["clients"].([]any)
-	var newClients []any
-	for client_index := range clients {
-		c := clients[client_index].(map[string]any)
-		if c["email"] == clientEmail {
-			c["totalGB"] = totalGB * 1024 * 1024 * 1024
-			c["updated_at"] = time.Now().Unix() * 1000
-			newClients = append(newClients, any(c))
-		}
-	}
-	settings["clients"] = newClients
-	modifiedSettings, err := json.MarshalIndent(settings, "", "  ")
-	if err != nil {
-		return false, err
-	}
-	inbound.Settings = string(modifiedSettings)
-	needRestart, err := s.UpdateInboundClient(inboundSvc, inbound, clientEmail)
-	return needRestart, err
-}
